% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{all_equal}
\alias{all_equal}
\title{Does \code{x} = \code{y}?}
\usage{
all_equal(x, y)
}
\arguments{
\item{x}{\code{vector}}

\item{y}{\code{vector}}
}
\value{
\code{lgl/chr} See \link[base]{all.equal}
}
\description{
Does \code{x} = \code{y}?
}
