% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{all_in}
\alias{all_in}
\title{Are all items in \code{x} in \code{y}?}
\usage{
all_in(x, y)
}
\arguments{
\item{x}{\code{vector}}

\item{y}{\code{vector}}
}
\value{
\code{lgl/chr} See \link[base]{all.equal}
}
\description{
Are all items in \code{x} in \code{y}?
}
