% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{as_js}
\alias{as_js}
\title{Preserve a string as JS/HTML (prevent translation of characters)}
\usage{
as_js(x)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{chr, HTML, JS_EVAL}
}
\description{
Preserve a string as JS/HTML (prevent translation of characters)
}
