% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environments_namespaces.R
\name{assign_global}
\alias{assign_global}
\title{Assign an object to the global environment}
\usage{
assign_global(x, nm = rlang::expr_deparse(rlang::enexpr(x)))
}
\arguments{
\item{x}{\code{obj} to assign}

\item{nm}{\code{chr} vector of the object and subsets to assign}
}
\value{
\code{x}
}
\description{
Assign an object to the global environment
}
\examples{
assign_global(FALSE, "test")
}
\seealso{
Other namespaces: 
\code{\link{assign_in_ns}()},
\code{\link{get_from_ns}()},
\code{\link{get_global}()},
\code{\link{pkg_ns}()},
\code{\link{unload_namespaces}()}
}
\concept{namespaces}
