% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{character_codes}
\alias{character_codes}
\title{Replace HTML Character codes with their character equivalent}
\usage{
character_codes(x, to_character = TRUE)
}
\arguments{
\item{x}{\code{chr} string(s) in which character codes should be replaced}

\item{to_character}{\code{lgl} Change character codes to character, if \code{FALSE} character symbols are changed to character codes}
}
\value{
\code{chr} with substitutions
}
\description{
See \code{?.character_codes} for conversion table. \strong{Note} that this will not translate Ampersand if converting from Namedcode to character because it will translate the Namedcodes themselves.
}
\examples{
character_codes("5 < 10")
character_codes("5 &lt; 10")
character_codes("5 &lt; 10", to_character = FALSE)
}
