% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\docType{data}
\name{col_type_hash}
\alias{col_type_hash}
\title{Index of column type conversions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.
}
\usage{
col_type_hash
}
\description{
Index of column type conversions
}
\keyword{datasets}
