% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{col_types}
\alias{col_types}
\title{Converts input to a specified type output}
\usage{
col_types(x, outtype = c("chr", "hud", "fun", "typ")[1])
}
\arguments{
\item{x}{\code{(vector/function)} One of:
\itemize{
\item{column}{ \code{(any)}}
\item{a type specification from HUD}{ \code{(character)}}
\item{a readr \verb{parse_*} function (See \link[readr]{parse_logical})}{ \code{(function)}}
\item{a readr type specification (See \link[readr]{cols})}{ \code{(character)}}
}}

\item{outtype}{\code{(character)} One of:
\itemize{
\item{\code{"chr"}}{ Returns the class as a readr abbreviation (See \link[readr]{cols})}
\item{\code{"hud"}}{ \code{(character)} a type specification from HUD}
\item{\code{"fun"}}{a readr \verb{parse_*} function (See \link[readr]{parse_logical})}{ \code{(function)}}
\item{\code{"type"}}{ \code{(character)} The R data class}
}}
}
\value{
See outtype
}
\description{
Given various inputs, provide a col_type specification in the format indicated by \code{outtype}
}
