% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_distance}
\alias{color_distance}
\title{Compute color distance}
\usage{
color_distance(x, y)
}
\arguments{
\item{x}{\code{num/chr} A Named vector of CSS hex or rgb/rgba color}

\item{y}{\code{num/chr} A Named vector of CSS hex or rgb/rgba color}
}
\value{
\code{num} A representation of the distance for comparison with other distances
}
\description{
Compute color distance
}
\examples{
color_distance(c(a = "rgba(111,96,140,1)"), c("#12B4D3", "green"))
}
