% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_css_color.R
\name{color_interpolate}
\alias{color_interpolate}
\title{Interpolate between two colors}
\usage{
color_interpolate(colors = c("#9A3324", "#016f90"), n, ...)
}
\arguments{
\item{colors}{\code{chr} of hex colors}

\item{n}{\code{int} of the number of colors expected}

\item{...}{
  Arguments passed on to \code{\link[grDevices:colorRamp]{grDevices::colorRamp}}
  \describe{
    \item{\code{bias}}{a positive number.  Higher values give more widely spaced
    colors at the high end.}
    \item{\code{space}}{a character string; interpolation in RGB or CIE Lab
    color spaces.}
    \item{\code{interpolate}}{use spline or linear interpolation.}
    \item{\code{alpha}}{logical: should alpha channel (opacity) values be
    returned?   It is an error to give a true value if
    \code{space} is specified.}
  }}
}
\value{
\code{chr} of hex values
}
\description{
Interpolate between two colors
}
\examples{
color_interpolate(n = 3)
}
