% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_luminance}
\alias{color_luminance}
\title{Find the luminance of a particular color, scaled 0-1}
\usage{
color_luminance(red, green, blue, alpha, ...)
}
\arguments{
\item{red, blue, green, alpha}{numeric vectors with values in \eqn{[0, M]}
    where \eqn{M} is \code{maxColorValue}.  When this is \code{255}, the
    \code{red}, \code{blue}, \code{green}, and \code{alpha}
    values are coerced to integers in \code{0:255} and the result is
    computed most efficiently.}

\item{...}{additional arguments}
}
\value{
\code{num}
}
\description{
Find the luminance of a particular color, scaled 0-1
}
\examples{
color_luminance(css_col2vec('white'))
}
