% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_match}
\alias{color_match}
\title{Match colors by visual distance}
\usage{
color_match(x, y)
}
\arguments{
\item{x}{\code{chr} CSS representations of colors that will be selected from as matches}
}
\value{
\code{tbl} Of matches for all of \code{x} with the associated distance
}
\description{
Helpful for pairing colors across light/dark palettes
}
