% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_text_by_luminance}
\alias{color_text_by_luminance}
\title{Set text color based on luminance}
\usage{
color_text_by_luminance(colors, text_light = "white", text_dark = "black")
}
\arguments{
\item{colors}{\code{chr} of css colors}

\item{text_light}{\code{chr} CSS color for light text}

\item{text_dark}{\code{chr} CSS color for dark text}
}
\value{
\code{chr} CSS text colors
}
\description{
Useful for applying one or another of text colors based on the luminance of a background
}
\examples{
color_text_by_luminance(c("white", "magenta", "red", "brown", "yellow"))
}
