% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{common_names}
\alias{common_names}
\title{Find the names in common}
\usage{
common_names(...)
}
\arguments{
\item{...}{\code{(objects)}}
}
\value{
\code{(character)} of the common names
}
\description{
Given named objects, find the names in common
}
