% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{comparison_inverse}
\alias{comparison_inverse}
\title{Convert a math comparator to it's inverse}
\usage{
comparison_inverse(x)
}
\arguments{
\item{x}{\code{chr/fun} Math comparator, one of <, >, <=, >=}
}
\value{
\code{chr/fun} the inverse comparator as the same class as \code{x}
}
\description{
Convert a math comparator to it's inverse
}
\examples{
comparison_inverse(">=")
}
