% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\docType{data}
\name{comparison_inverse_key}
\alias{comparison_inverse_key}
\title{Math comparison comparator inverse key}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
comparison_inverse_key
}
\description{
Math comparison comparator inverse key
}
\seealso{
Other math: 
\code{\link{comparison_inverse}()},
\code{\link{comparison_key}},
\code{\link{evens}()},
\code{\link{interpolate}()},
\code{\link{odds}()},
\code{\link{str_comparison}()}
}
\concept{math}
\keyword{datasets}
