% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\docType{data}
\name{comparison_key}
\alias{comparison_key}
\title{Math comparison comparator to plain english key}
\format{
An object of class \code{character} of length 7.
}
\usage{
comparison_key
}
\description{
Math comparison comparator to plain english key
}
\seealso{
Other math: 
\code{\link{comparison_inverse}()},
\code{\link{comparison_inverse_key}},
\code{\link{evens}()},
\code{\link{interpolate}()},
\code{\link{odds}()},
\code{\link{str_comparison}()}
}
\concept{math}
\keyword{datasets}
