% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{concat_rows}
\alias{concat_rows}
\title{Concatenate row values in a poorly scraped table}
\usage{
concat_rows(.data, col_to_check = 1)
}
\arguments{
\item{.data}{\code{tbl} Of data with empty rows}

\item{col_to_check}{\code{num} The column with rows populated into which subsequent rows will be collapsed.}
}
\value{
\code{tbl}
}
\description{
Concatenate row values in a poorly scraped table
}
