% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environments_namespaces.R
\name{create_simple_get_function}
\alias{create_simple_get_function}
\title{Create a function that creates an object and assigns it to a namespace the first time it's called and subsequently retrieves it from the namespace thereafter.}
\usage{
create_simple_get_function(obj_nm, env_expr, call_expr)
}
\arguments{
\item{obj_nm}{\code{chr} Name of the object}

\item{call_expr}{\code{expr} The code used to construct the object if the object hasn't already been constructed}

\item{ns_chr}{\code{expr/chr} name of the namespace to assign the object to, or an expression that returns the environment to assign to}
}
\value{
\code{chr} The function at the console for copy/paste
}
\description{
Useful when the object depends on a long running task such as a database query.
}
\examples{
create_simple_get_function("mt_cars", .GlobalEnv, dplyr::mutate(mtcars, cyl = as.character(cyl)))
}
