% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{creds_to_renviron}
\alias{creds_to_renviron}
\title{Write named credentials to .Renviron file}
\usage{
creds_to_renviron(
  ...,
  scope = c("user", "project"),
  overwrite = FALSE,
  proj_dir = "."
)
}
\arguments{
\item{...}{named keys to write}

\item{scope}{Edit globally for the current \strong{user}, or locally for the
current \strong{project}}

\item{overwrite}{\code{(lgl)} should an existing key be overwritten. \strong{Default: \code{FALSE}}}

\item{proj_dir}{\code{(chr)} project directory to write credentials to}
}
\value{
success message if a value is written
}
\description{
Writes key pairs to .Renviron and adds .Renviron to \emph{.gitignore} if not already there.
}
