% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{dep_read}
\alias{dep_read}
\alias{object_read}
\title{Read a dependency from file}
\usage{
dep_read(filename, ...)

object_read(filename, ...)
}
\arguments{
\item{filename}{\code{(chr)} path to the file}

\item{...}{Passed on to read function. See \link[UU]{file_fn} for which function will be used based on the filename}
}
\value{
file contents
}
\description{
Read a dependency from file
}
\seealso{
Other file IO: 
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
