% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{dir_fn}
\alias{dir_fn}
\title{Create a directory path pointing function}
\usage{
dir_fn(base_dir)
}
\arguments{
\item{base_dir}{\code{(chr)} the base directory to which the path should point}
}
\value{
\code{(fun)} with pointing function using \link[fs]{path}
}
\description{
Create a directory path pointing function
}
\examples{
dir_fn("data")("random_data", "file", ext = "txt")
}
