% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\docType{data}
\name{.character_codes}
\alias{.character_codes}
\title{Character code conversion table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 4 columns.
}
\usage{
.character_codes
}
\description{
Character code conversion table
}
\keyword{datasets}
