% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_navigation.R
\docType{data}
\name{.file}
\alias{.file}
\title{All the files that can be navigated to with \code{file} if they exist}
\format{
An object of class \code{character} of length 12.
}
\usage{
.file
}
\description{
All the files that can be navigated to with \code{file} if they exist
}
\seealso{
Other file navigation: 
\code{\link{file}},
\code{\link{folder}}
}
\concept{file navigation}
\keyword{datasets}
