% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\docType{data}
\name{.time_factor}
\alias{.time_factor}
\title{Timespans as factor}
\format{
An object of class \code{ordered} (inherits from \code{factor}) of length 7.
}
\usage{
.time_factor
}
\description{
Timespans as factor
}
\keyword{datasets}
