% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{evens}
\alias{evens}
\title{Get even numbers}
\usage{
evens(x)
}
\arguments{
\item{x}{\code{int}}
}
\description{
Get even numbers
}
\examples{
evens(1:10)
}
\seealso{
Other math: 
\code{\link{comparison_inverse}()},
\code{\link{comparison_inverse_key}},
\code{\link{comparison_key}},
\code{\link{interpolate}()},
\code{\link{odds}()},
\code{\link{str_comparison}()}
}
\concept{math}
