% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\name{excel_date}
\alias{excel_date}
\title{Convert Excel character date representation to a Date}
\usage{
excel_date(.x)
}
\arguments{
\item{.x}{\code{chr} representation of Dates from Excel, usually a 5 digit number}
}
\value{
\code{Date}
}
\description{
Convert Excel character date representation to a Date
}
\examples{
excel_date(c("44567", "44568"))
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{time_elapsed}()},
\code{\link{week_factor}()}
}
\concept{time}
