% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{ext}
\alias{ext}
\title{Extract the file extensions from a filepath}
\usage{
ext(path)
}
\arguments{
\item{path}{\code{(character)} path}
}
\value{
\code{(character)} with the extensions
}
\description{
Given a path, extract the file extension
}
