% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{file_fn}
\alias{file_fn}
\title{Return the appropriate function for reading the specified path/extension}
\usage{
file_fn(x, write = FALSE)
}
\arguments{
\item{x}{\code{(character)} The extension name or the path to the file}

\item{write}{\code{(logical)} Return the writing function? \strong{Default} \code{FALSE} to return the reading function}
}
\value{
\code{(function)}
}
\description{
Return the appropriate function for reading the specified path/extension
}
\examples{
file_fn("csv")
file_fn("csv", write = TRUE)
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{load_obj}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
