% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{file_io}
\alias{file_io}
\title{Provide the appropriate file read/write function}
\usage{
file_io(x, path, mkpath = TRUE)
}
\arguments{
\item{object}{to determine the appropriate function for writing to disk}

\item{ext}{\code{(character)} file extension to determine the appropriate reading function}
}
\value{
\code{(function)}
}
\description{
Return the appropriate read or write function given an object or \code{ext}ension as a character
}
