% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{file_io_ext}
\alias{file_io_ext}
\title{Provide the appropriate file extension for a given object}
\usage{
file_io_ext(object)
}
\arguments{
\item{object}{to determine the appropriate function for writing to disk}
}
\value{
\code{(character)}
}
\description{
Provide the appropriate file extension for a given object
}
