% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date_time.R
\name{file_timestamp}
\alias{file_timestamp}
\title{Make a file path compliant ISO8601 timestamp}
\usage{
file_timestamp(x = Sys.time(), tzone = "")
}
\arguments{
\item{x}{\code{Datetime}}

\item{tzone}{a character string containing the time zone to convert to. R
must recognize the name contained in the string as a time zone on your
system.}
}
\value{
\code{chr} of timestamp
}
\description{
Make a file path compliant ISO8601 timestamp
}
\examples{
fs::path(paste0("file_", file_timestamp()), ext = "zip")
}
