% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{filter_to}
\alias{filter_to}
\title{Change or apply filters to output type}
\usage{
filter_to(filters, out_type = character(), .data)
}
\arguments{
\item{filters}{\code{list} of named filters to use. Each filter is named by the column name with the range as a length two numeric vector}

\item{out_type}{\code{obj} of desired output type}

\item{.data}{\code{tbl} data to use if \code{out_type = logical()/data.frame()}}
}
\value{
\code{obj} of same class as out_type
}
\description{
Useful in concert with axis brushing
}
\examples{
f <- list(wt = c(1,5))
d <- data.frame(wt = 1:10)
filter_to(f)
filter_to(f, out_type = data.frame(), .data = d)
}
