% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{find_by_class}
\alias{find_by_class}
\title{Find an object by it's class}
\usage{
find_by_class(class, e = rlang::caller_env())
}
\arguments{
\item{\code{(environment)}}{The environment to search}

\item{\code{(class)}}{The class to search for}
}
\description{
Find an object by it's class
}
