% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{fml_list}
\alias{fml_list}
\title{Print function formals as a list}
\usage{
fml_list(f, paired = TRUE)
}
\arguments{
\item{f}{\code{fun}}
}
\value{
\code{msg}
}
\description{
Print function formals as a list
}
\examples{
fml_list(base::apply)
}
