% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{fn_name}
\alias{fn_name}
\title{Retrieve the function name}
\usage{
fn_name(fn)
}
\arguments{
\item{fn}{\coe{(function)} for which to retrieve the name}
}
\value{
\code{(character)} of the functions name
}
\description{
Sometimes a function is passed down the call stack and it's name is unknown. This function finds the name without having to pass it down the call stack as an argument.
}
