% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_navigation.R
\docType{data}
\name{folder}
\alias{folder}
\title{Go to a specified folder}
\format{
An object of class \code{list} of length 4.
}
\usage{
folder
}
\arguments{
\item{path}{\code{chr} to go to. Each method has a preset default based on the function name, but any method can be used to go to any folder specified.}
}
\description{
Go to a specified folder
}
\seealso{
Other file navigation: 
\code{\link{.file}},
\code{\link{file}}
}
\concept{file navigation}
\keyword{datasets}
