% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{formula_make}
\alias{formula_make}
\title{Create a formula given predictors and a label (response variable)}
\usage{
formula_make(predictors, label = "response")
}
\arguments{
\item{predictors}{\code{chr} Column names used for prediction}

\item{label}{\code{chr} Column name of response variable}
}
\value{
\code{formula}
}
\description{
Create a formula given predictors and a label (response variable)
}
\examples{
formula_make(predictors = c("max2yrlfflow", "min2yrlfflow"))
}
