% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{fun_docs_table}
\alias{fun_docs_table}
\title{Create a table of functions and their uses}
\usage{
fun_docs_table(package = pkgload::pkg_name())
}
\arguments{
\item{package}{\code{chr} package name}
}
\value{
\code{shiny.tag}
}
\description{
Create a table of functions and their uses
}
