% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_helpers.R
\name{get_from_ns}
\alias{get_from_ns}
\title{Get an object from a namespace}
\usage{
get_from_ns(nm = c("active", "state")[1], .env = .GlobalEnv)
}
\arguments{
\item{nm}{\code{chr} name of object to retrieve. Current are \code{active} & \code{state}.}

\item{.env}{\code{env} from which to retrieve the object
\itemize{
\item{\code{active}}{ tracks app details for internal use}
\item{\code{state}}{ tracks user-specified variables to be saved/exported}
}}
}
\value{
\code{obj}
}
\description{
Get an object from a namespace
}
