% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environments_namespaces.R
\name{get_global}
\alias{get_global}
\title{Get an object from the global environment}
\usage{
get_global(global = "active")
}
\value{
\code{obj}
}
\description{
Get an object from the global environment
}
\examples{
get_global(".Last.value")
}
