% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{get_package_fns}
\alias{get_package_fns}
\title{Get the names of all exported functions in a package}
\usage{
get_package_fns(x, all.names = FALSE)
}
\arguments{
\item{x}{\code{(character)} Package name}

\item{all.names}{\code{(logical)} Include names that begin with characters \code{.} \verb{_} etc}
}
\value{
\code{(character)}
}
\description{
Get the names of all exported functions in a package
}
\examples{
get_package_fns("dplyr")
}
