% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gitignore_alphabetize}
\alias{gitignore_alphabetize}
\title{Alphabetize the \emph{.gitignore} file}
\usage{
gitignore_alphabetize(gitignore = ".gitignore")
}
\arguments{
\item{gitignore}{\code{chr} path to gitignore}
}
\value{
\code{None} overwrites the file with entries in alphabetical order
}
\description{
Alphabetize the \emph{.gitignore} file
}
