% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{glue_js}
\alias{glue_js}
\title{Create a JS string with glue insertions
glue \verb{.open = !@} & \verb{.close = @#}}
\usage{
glue_js(
  js,
  as_chr = FALSE,
  e = rlang::caller_env(),
  .open = "*{",
  .close = "}*"
)
}
\arguments{
\item{js}{\code{chr} JS code to \link[glue]{glue}}

\item{as_chr}{\code{lgl} Whether to use \code{\link[UU]{as_js}} on the output \code{FALSE} or \link[base]{as.character} \code{TRUE}. \strong{Default FALSE}}

\item{e}{\code{env} calling environment}

\item{.open}{[\code{character(1)}: \sQuote{\\{}]\cr The opening delimiter. Doubling the
full delimiter escapes it.}

\item{.close}{[\code{character(1)}: \sQuote{\\}}]\cr The closing delimiter. Doubling the
full delimiter escapes it.}
}
\value{
\code{chr}
}
\description{
Create a JS string with glue insertions
glue \verb{.open = !@} & \verb{.close = @#}
}
\examples{
glue_js("$(document).ready(() => {let x = *{tolower(FALSE)}*)")
}
