% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gmessage}
\alias{gmessage}
\title{Custom message
Message using \link[cli]{format_message}}
\usage{
gmessage(e = rlang::caller_env())
}
\description{
Custom message
Message using \link[cli]{format_message}
}
