% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gmsg}
\alias{gmsg}
\title{Custom message
Message using \link[cli]{format_message} & \link[cli]{cat_line}}
\usage{
gmsg(msg, e = rlang::caller_env())
}
\description{
Custom message
Message using \link[cli]{format_message} & \link[cli]{cat_line}
}
