% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infixes.R
\name{\%allin\%}
\alias{\%allin\%}
\title{Are all lhs values in rhs?}
\usage{
lhs \%allin\% rhs
}
\description{
Are all lhs values in rhs?
}
\examples{
2:5 \%allin\% 1:10
5:11 \%allin\% 1:10
}
\seealso{
Other infixes: 
\code{\link{\%nin\%}()},
\code{\link{\%|0|\%}()},
\code{\link{\%|\%}()},
\code{\link{\%|legit|\%}()},
\code{\link{\%|try|\%}()},
\code{\link{\%|zchar|\%}()}
}
\concept{infixes}
