% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infixes.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Are lhs values absent from set on rhs?}
\usage{
... \%not_in\% NA
}
\description{
Are lhs values absent from set on rhs?
}
\examples{
1 \%not_in\% 1:10
}
\seealso{
is_legit
}
\author{
Think.fr
}
