% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infixes.R
\name{\%|\%}
\alias{\%|\%}
\title{Replace NA values in LHS with RHS}
\usage{
lhs \%|\% rhs
}
\arguments{
\item{lhs}{\code{vctr} of values possibly containing \code{NA} on which replacement will be performed}

\item{rhs}{\code{vctr} of values length 1 or the same length with which to replace \code{NA}}
}
\value{
\code{vctr} with class according to R's coercion rules.
}
\description{
Replace NA values in LHS with RHS
}
\examples{
c(4, NA, 4) \%|\% 3
c(4, NA, 4) \%|\% 3L
class(c(4, NA, 4, NA) \%|\% 3L)
cclass((NA, NA, NA, NA) \%|\% 3L)
c(4, NA, 4, NA) \%|\% c(2,2,2,3)
}
