% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infixes.R
\name{\%|legit|\%}
\alias{\%|legit|\%}
\title{If legit lhs, else rhs}
\usage{
lhs \%|legit|\% rhs
}
\value{
If legit lhs else rhs
}
\description{
If legit lhs, else rhs
}
\examples{
(100 / NA) \%|legit|\% 4
list(a = 5)$a \%|legit|\% 4
list(a = 5)$b \%|legit|\% 4
}
\seealso{
is_legit
}
