% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infixes.R
\name{\%|nzchar|\%}
\alias{\%|nzchar|\%}
\title{Replace zero-length character strings with right hand side}
\usage{
lhs \%|nzchar|\% rhs
}
\arguments{
\item{lhs}{\code{chr}}

\item{rhs}{\code{chr}}
}
\value{
\code{chr}
}
\description{
Replace zero-length character strings with right hand side
}
\examples{
c("a" , "", "c", "") \%|nzchar|\% "b"
}
