% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infixes.R
\name{\%|zchar|\%}
\alias{\%|zchar|\%}
\title{Replace zero-length character strings with right hand side}
\usage{
lhs \%|zchar|\% rhs
}
\arguments{
\item{lhs}{\code{chr}}

\item{rhs}{\code{chr}}
}
\value{
\code{chr}
}
\description{
Replace zero-length character strings with right hand side
}
\examples{
c("a" , "", "c", "") \%|zchar|\% "b"
}
\seealso{
Other infixes: 
\code{\link{\%allin\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%|0|\%}()},
\code{\link{\%|\%}()},
\code{\link{\%|legit|\%}()},
\code{\link{\%|try|\%}()}
}
\concept{infixes}
