% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gwarn}
\alias{gwarn}
\title{Custom warning message}
\usage{
gwarn(
  message = NULL,
  class = NULL,
  ...,
  .frequency = c("always", "regularly", "once"),
  e = rlang::caller_env()
)
}
\arguments{
\item{message}{\code{(chr)} The message to display, formatted as a bulleted list. The first element is displayed as an alert bullet prefixed with ! by default. Elements named "*", "i", and "x" are formatted as regular, info, and cross bullets respectively. See \link[rlang]{topic-condition-formatting} for more about bulleted messaging.}

\item{class}{\code{(class)} Subclass of the condition}

\item{.frequency}{\code{(chr)} How frequently should the warning or message be displayed? By default ("always") it is displayed at each time. If "regularly", it is displayed once every 8 hours. If "once", it is displayed once per session.}

\item{e}{\code{(environment)} calling environment. Passed to \code{glue} for making the message}
}
\description{
Throw \link[rlang]{warn} with \link[cli]{format_warning}
}
