% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_opts.R
\name{if_debug}
\alias{if_debug}
\title{Run expressions only when option \code{use_debug = TRUE}}
\usage{
if_debug(...)
}
\arguments{
\item{...}{\code{expr} to evaluate}
}
\description{
Run expressions only when option \code{use_debug = TRUE}
}
\seealso{
Other options: 
\code{\link{opts}},
\code{\link{toggle}},
\code{\link{write_opts}()}
}
\concept{options}
