% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{ignore_files}
\alias{ignore_files}
\title{Add lines to \emph{.gitignore}}
\usage{
ignore_files(lines, directory = ".", ignore_file = ".gitignore")
}
\arguments{
\item{lines}{\code{chr}}

\item{directory}{\code{(chr)} directory path to \code{ignore_file}}

\item{ignore_file}{\code{chr} filename holding ignores to be modified. Default \emph{.gitignore}}
}
\value{
\code{informative messages}
}
\description{
Add lines to \emph{.gitignore}
}
