% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{increment}
\alias{increment}
\title{Increment an in-place counter}
\usage{
increment(x, e = rlang::caller_env())
}
\arguments{
\item{x}{\code{num/reactiveVal/reactiveValues} Any numeric to be incremented}

\item{e}{\code{env} Calling environment}
}
\value{
\code{none} Increments the counter in the parent environment (modifies in place)
}
\description{
Increment an in-place counter
}
\examples{
x <- 1
increment(x)
x
}
