% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\docType{data}
\name{inequality_key}
\alias{inequality_key}
\title{Inequality conversion key}
\format{
An object of class \code{character} of length 7.
}
\usage{
inequality_key
}
\description{
Inequality conversion key
}
\keyword{datasets}
