% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_setup.R
\name{install_yogat3ch}
\alias{install_yogat3ch}
\title{Install packages from the yogat3ch repo}
\usage{
install_yogat3ch(pkg, ..., to_desc = TRUE, snapshot = TRUE)
}
\arguments{
\item{pkg}{\code{chr} package names \strong{Required}}

\item{...}{\code{args} passed on to \link[base]{install.packages} if no \code{remote} supplied or \link[remotes]{install_github} if \code{remote} supplied}

\item{to_desc}{\code{lgl} Add the package dependency to the \emph{DESCRIPTION} file?}

\item{snapshot}{\code{lgl} Run \code{renv::\link[renv]{snapshot}}}
}
\description{
Install packages from the yogat3ch repo
}
