% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_package_dev.R
\name{installed_to_description}
\alias{installed_to_description}
\title{Changes installed package version relationships to \code{==} from the \code{usethis} default of \code{>=}}
\usage{
installed_to_description(pkg, write_to_desc = TRUE)
}
\arguments{
\item{pkg}{\code{chr} Name of package}

\item{write_to_desc}{\code{lgl} Whether to write the packages with updated}
}
\value{
\code{None} Either prints to the console or modifies the DESCRIPTION in place
}
\description{
Changes installed package version relationships to \code{==} from the \code{usethis} default of \code{>=}
}
