% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{interpolate}
\alias{interpolate}
\title{Simple interpolate between two numbers}
\usage{
interpolate(start, end, n)
}
\arguments{
\item{start}{\code{num} first number}

\item{end}{\code{num} last number}

\item{n}{\code{int} total numbers in output vector}
}
\value{
\code{num} vector with length equivalent to n
}
\description{
Simple interpolate between two numbers
}
\examples{
interpolate(0, 6, 6)
}
\seealso{
Other math: 
\code{\link{comparison_inverse}()},
\code{\link{comparison_inverse_key}},
\code{\link{comparison_key}},
\code{\link{evens}()},
\code{\link{odds}()},
\code{\link{str_comparison}()}
}
\concept{math}
