% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditionals.R
\name{is_error}
\alias{is_error}
\title{Is object an error}
\usage{
is_error(x)
}
\arguments{
\item{x}{\code{obj}}
}
\value{
\code{lgl}
}
\description{
Is object an error
}
\examples{
is_error(try(stop()))
}
\seealso{
Other conditionals: 
\code{\link{is_legit}()},
\code{\link{is_project}()},
\code{\link{larger}()},
\code{\link{most}()},
\code{\link{same}()},
\code{\link{zchar}()}
}
\concept{conditionals}
