% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{is_filepath}
\alias{is_filepath}
\title{Is path a file path}
\usage{
is_filepath(path)
}
\arguments{
\item{path}{\code{(character)} path}
}
\value{
\code{(logical)}
}
\description{
Given a path, is it a filepath?
}
