% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_package_dev.R
\name{is_package_dev}
\alias{is_package_dev}
\title{Is package in development or installed}
\usage{
is_package_dev(pkg_nm = pkg_name())
}
\arguments{
\item{pkg_nm}{\code{chr} name of package}
}
\value{
\code{lgl}
}
\description{
Is package in development or installed
}
\seealso{
Other package dev: 
\code{\link{get_package_fns}()},
\code{\link{is_package}()},
\code{\link{need_pkg}()},
\code{\link{pkg_chr_split_comma}()}
}
\concept{package dev}
