% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditionals.R
\name{is_project}
\alias{is_project}
\title{Is Session in a Project?}
\usage{
is_project()
}
\value{
\code{lgl}
}
\description{
Is Session in a Project?
}
\examples{
is_project()
}
\seealso{
Other conditionals: 
\code{\link{is_error}()},
\code{\link{is_legit}()},
\code{\link{larger}()},
\code{\link{most}()},
\code{\link{same}()},
\code{\link{zchar}()}
}
\concept{conditionals}
