% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{join_check}
\alias{join_check}
\title{Detect possible duplicates of rows or columns after a join}
\usage{
join_check(after, before, halt_fn = rlang::warn)
}
\arguments{
\item{after}{\code{(data.frame)} after the join}

\item{before}{\code{(data.frame)} from before the join \strong{Optional} but required for row comparison.}

\item{halt_fn}{\code{(function)} to notify, default \link[rlang]{warn}.}
}
\description{
Detect possible duplicates of rows or columns after a join
}
\examples{
a = data.frame(a = c(1, 2, 3, 4, 3, 5), b = 1:6)
b = data.frame(a = c(1, 2, 3, 4, 5), c = letters[1:5])
after <- dplyr::left_join(a, b)
join_check(b, after, halt_fn = message)
}
\seealso{
\code{\link[rlang:abort]{rlang::abort()}} \code{\link[base:message]{base::message()}}
}
