% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{json_validate}
\alias{json_validate}
\title{Is JSON file or text valid?}
\usage{
json_validate(x)
}
\arguments{
\item{x}{\code{chr} JSON string or file path}
}
\value{
\code{lgl} Whether JSON is valid
}
\description{
Is JSON file or text valid?
}
\examples{
json_validate('{"done" : 1}')
}
