% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{key_out}
\alias{key_out}
\title{Handle different output type requests for \code{match_df}}
\usage{
key_out(x, keys, out)
}
\description{
Handle different output type requests for \code{match_df}
}
