% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{key_pairs_duplicated}
\alias{key_pairs_duplicated}
\title{Find duplicates in key pairs}
\usage{
key_pairs_duplicated(x, fromLast = TRUE)
}
\arguments{
\item{x}{\code{chr/Dlist} Either character strings of keypairs or a Dlist returned by \link[base]{Sys.getenv}}
}
\value{
\code{lgl}
}
\description{
Find duplicates in key pairs
}
\examples{
key_pairs_duplicated(Sys.getenv())
}
