% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_setup.R
\name{key_pairs_text}
\alias{key_pairs_text}
\title{Make key-pairs from a named character vector}
\usage{
key_pairs_text(Sys.getenv())
}
\arguments{
\item{x}{\code{chr} named}
}
\value{
\code{chr}
}
\description{
Make key-pairs from a named character vector
}
\seealso{
Other project setup: 
\code{\link{creds_to_renviron}()},
\code{\link{fun_docs_table}()},
\code{\link{ignore_files}()},
\code{\link{install_remote}()},
\code{\link{key_pairs_duplicated}()},
\code{\link{use_UU_reimports}()},
\code{\link{use_reimport}()},
\code{\link{write_to_rprofile}()}
}
\concept{project setup}
