% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{last_updated}
\alias{last_updated}
\title{Gather last updated times for on-disk files}
\usage{
last_updated(x, path = "data")
}
\arguments{
\item{x}{\code{(chr)} file path to check last updated time}

\item{path}{\code{(chr)} directory path in which to check all files (alternative to \code{x})}
}
\value{
\code{(POSIXct)} Last modified time
}
\description{
Check the last modified time for the hud_exports specified
}
