% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vectors.R
\name{len_unique}
\alias{len_unique}
\title{The length of unique values in a vector}
\usage{
len_unique(x)
}
\arguments{
\item{x}{\code{vctr}}
}
\value{
\code{dbl}
}
\description{
The length of unique values in a vector
}
\examples{
len_unique(c(1,2,1))
}
\seealso{
Other vectors: 
\code{\link{names_values_switch}()},
\code{\link{rle_df}()},
\code{\link{rle_groups}()},
\code{\link{rle_seq}()},
\code{\link{sort_by_names}()},
\code{\link{true_names}()},
\code{\link{unify_vec_preserve_order}()},
\code{\link{unique_with_names}()},
\code{\link{vlookup_from_ref}()},
\code{\link{zchar_remove}()}
}
\concept{vectors}
