% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{list.files2}
\alias{list.files2}
\title{List full file paths with the file name as the name}
\usage{
list.files2(path)
}
\arguments{
\item{path}{\code{(character)} dir path to files}
}
\value{
\code{(named character)}
}
\description{
List full file paths with the file name as the name
}
\examples{
list.files2("~")
}
