% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{list_rename}
\alias{list_rename}
\title{Rename a list}
\usage{
list_rename(.data, ...)
}
\arguments{
\item{.data}{\code{list} To be renamed}

\item{...}{\code{named arguments} in the form new_name = old_name (unquoted)}
}
\value{
\code{list}
}
\description{
From \href{https://stackoverflow.com/users/6646912/krassowski}{krassowski} on SO \href{https://stackoverflow.com/a/73621060/2675597}{link}
}
\examples{
my_list = list(a=1, b=2, c=3)
list_rename(my_list, x=a, y=b)
}
