% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{load_obj}
\alias{load_obj}
\title{Load an R object from a file}
\usage{
load_obj(file)
}
\arguments{
\item{file}{\code{character} A character string specifying the file path.}
}
\value{
The loaded R object.
}
\description{
This function loads an R object from a file into the global environment or a new environment.
}
\examples{
# Load an R object from a file
obj <- load_obj("path/to/file.RData")

}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
