% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_css_color.R
\name{luminance_filter}
\alias{luminance_filter}
\title{Filter colors based on a luminance threshold}
\usage{
luminance_filter(colors, dark_mode, luminance_threshold = 0.5)
}
\arguments{
\item{colors}{\code{chr} color vector}

\item{dark_mode}{\code{lgl} colors will be \code{>=} if \code{TRUE}, otherwise \code{<=} if \code{FALSE}}

\item{luminance_threshold}{\code{lgl} 0-1 luminance threshold to filter}
}
\value{
\code{chr} vector of colors
}
\description{
Filter colors based on a luminance threshold
}
\examples{
x <- tibble::tribble(
~CSS.Name, ~Hex.Code, ~Decimal.Code,
"Black",        "#000000",            "(0,0,0)",
"White",        "#FFFFFF",      "(255,255,255)",
"Red",        "#FF0000",          "(255,0,0)",
"Lime",        "#00FF00",          "(0,255,0)",
"Blue",        "#0000FF",          "(0,0,255)",
"Yellow",        "#FFFF00",        "(255,255,0)",
"Cyan / Aqua",        "#00FFFF",        "(0,255,255)",
"Magenta / Fuchsia",        "#FF00FF",        "(255,0,255)",
"Silver",        "#C0C0C0",      "(192,192,192)",
"Gray",        "#808080",      "(128,128,128)",
"Maroon",        "#800000",          "(128,0,0)",
"Olive",        "#808000",        "(128,128,0)",
"Green",        "#008000",          "(0,128,0)",
"Purple",        "#800080",        "(128,0,128)",
"Teal",        "#008080",        "(0,128,128)",
"Navy",        "#000080",          "(0,0,128)"
)
luminance_filter(x$Hex.Code, TRUE)
luminance_filter(x$Hex.Code, FALSE)
}
\seealso{
Other color: 
\code{\link{color_cycle}()},
\code{\link{color_distance}()},
\code{\link{color_luminance}()},
\code{\link{color_match}()},
\code{\link{color_rgb_table}()},
\code{\link{color_separate}()},
\code{\link{color_text_by_luminance}()},
\code{\link{colors2css}()},
\code{\link{css_col2vec}()},
\code{\link{rgb2hex}()}
}
\concept{color}
