% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{make_names}
\alias{make_names}
\title{Make a file path name with underscores}
\usage{
make_names(x)
}
\arguments{
\item{\code{(character)}}{file path}
}
\description{
Make a file path name with underscores
}
