% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{map_class}
\alias{map_class}
\title{Match the classes of one object to that of another object}
\usage{
map_class(x, y)
}
\arguments{
\item{x}{\code{(object)} object to be matched}

\item{y}{\code{(object)} object to be coerced}
}
\value{
\code{y} with class types matching that of \code{x}
}
\description{
Match the classes of one object to that of another object
}
