% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{max}
\alias{max}
\title{An alternative to \link[base]{max} that preserves names}
\usage{
max(x)
}
\arguments{
\item{x}{\code{vec}}
}
\value{
\code{atomic} returns the largest element in the vector
}
\description{
An alternative to \link[base]{max} that preserves names
}
\examples{
max(c(a = 1, b = 2))
}
