% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{max2}
\alias{max2}
\title{An alternative to \link[base]{max} that preserves names}
\usage{
max2(x)
}
\arguments{
\item{x}{\code{vec}}
}
\value{
\code{atomic} returns the largest element in the vector
}
\description{
An alternative to \link[base]{max} that preserves names
}
\examples{
max2(c(a = 1, b = 2))
}
\seealso{
Other statistics: 
\code{\link{smode}()}
}
\concept{statistics}
