% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{mkpath}
\alias{mkpath}
\title{Construct a path}
\usage{
mkpath(path)
}
\arguments{
\item{path}{\code{(character)} path}
}
\value{
\code{(logical)}
}
\description{
Given a path, construct it if it does not exist.
}
