% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date_time.R
\name{month_factor}
\alias{month_factor}
\title{Month as factor/numeric}
\usage{
month_factor(x, label = FALSE, abbr = TRUE)
}
\arguments{
\item{x}{\code{chr/Date/POSIXt} Names of months as case insensitive character vector, if using abbreviations, they must be of consistent length. Otherwise, a \code{Date} or \code{POSIXt}.}
}
\value{
\code{factor} If \code{x} is provided, the month as a factor. If \code{x} is not provided, an ordered factor of the months
}
\description{
Month as factor/numeric
}
\examples{
month_factor()
month_factor(c('Se', 'No'))
month_factor(c('Se', 'No'), label = TRUE)
month_factor(seq(lubridate::floor_date(Sys.time(), "year"), Sys.time(), by = "day"), label = TRUE, abbr = FALSE)
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{excel_date}()},
\code{\link{season_factor}()},
\code{\link{time_aggregates}},
\code{\link{time_difftimes}},
\code{\link{time_elapsed}()},
\code{\link{time_factor}()},
\code{\link{timespan}()},
\code{\link{week_factor}()}
}
\concept{time}
