% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{most}
\alias{most}
\title{Are most values TRUE}
\usage{
most(x)
}
\arguments{
\item{x}{\code{lgl}}
}
\value{
\code{lgl}
}
\description{
IF more than half the values are TRUE, returns TRUE
}
\examples{
most(c(TRUE,TRUE,FALSE))
most(c(TRUE,FALSE,FALSE))
}
\seealso{
Other conditionals: 
\code{\link{all_equal}()},
\code{\link{is_error}()},
\code{\link{is_legit}()},
\code{\link{is_project}()},
\code{\link{larger}()}
}
\concept{conditionals}
