% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{move_files_to_folder}
\alias{move_files_to_folder}
\title{Move all files to a folder}
\usage{
move_files_to_folder(
  files = UU::list.files2("inst/app/www", pattern = "\\\\.js$", include.dirs = FALSE),
  folder = dirs$js()
)
}
\arguments{
\item{files}{\code{chr} of files to move}

\item{folder}{\code{chr} folder to move them to}
}
\description{
Move all files to a folder
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
