% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_navigation.R
\name{move_js_to_folder}
\alias{move_js_to_folder}
\title{Move all js files to js folder}
\usage{
move_js_to_folder(
  files = UU::list.files2("inst/app/www", pattern = "\\\\.js$", include.dirs = FALSE),
  folder = dirs$js()
)
}
\description{
Move all js files to js folder
}
