% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{names_values_switch}
\alias{names_values_switch}
\title{Switch the names and the values of a vector}
\usage{
names_values_switch(x)
}
\arguments{
\item{x}{\code{named object}}
}
\value{
\code{obj}
}
\description{
Switch the names and the values of a vector
}
\examples{
names_values_switch(c(a = 1, b = 2))
}
\seealso{
Other vectors: 
\code{\link{vlookup_from_ref}()},
\code{\link{vlookup}()}
}
\concept{vectors}
