% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{need_pkg}
\alias{need_pkg}
\title{Get a function from a package, abort if package not installed.}
\usage{
need_pkg(pkg, fn)
}
\arguments{
\item{pkg}{\code{chr} package}

\item{fn}{\code{fn} function name}
}
\value{
\code{fun}
}
\description{
Get a function from a package, abort if package not installed.
}
\examples{
need_pkg("utils", "recover")
}
\concept{development}
