% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{needs_update}
\alias{needs_update}
\title{Check if files need to be updated}
\usage{
needs_update(
  x,
  path = FALSE,
  threshold = lubridate::floor_date(Sys.time(), "day")
)
}
\arguments{
\item{x}{\code{(chr)} file path to check last updated time}

\item{path}{\code{(lgl)} whether x is a path and all files should be checked}

\item{threshold}{The threshold time. If files have last modified times less than this time, they will be marked as needing an update.}
}
\value{
\code{(tbl)} with columns:
\itemize{
\item{\code{full_path}}{ The full path to the file(s)}
\item{\code{basename}}{ The file(s) basename}
\item{\code{last_updated}}{ The last updated time}
\item{\code{threshold}}{ The threshold time for comparison}
\item{\code{needs_update}}{ logical as to whether the file should be updated}
}
}
\description{
Check if files need to be updated
}
