% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{num2str}
\alias{num2str}
\title{Convert numeric value to a string abbreviation with K, M, B for Thousand, Million & Billion}
\usage{
num2str(n)
}
\arguments{
\item{n}{\code{num}}
}
\value{
\code{chr}
}
\description{
Convert numeric value to a string abbreviation with K, M, B for Thousand, Million & Billion
}
\examples{
num2str(10000)
}
