% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\docType{data}
\name{num_chr_suffi}
\alias{num_chr_suffi}
\title{Abbreviations of numeric magnitude}
\format{
An object of class \code{character} of length 4.
}
\usage{
num_chr_suffi
}
\description{
Abbreviations of numeric magnitude
}
\keyword{datasets}
