% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{num_suf}
\alias{num_suf}
\title{What is the human-readable suffix for a number}
\usage{
num_suf(n)
}
\arguments{
\item{n}{\code{num/chr} Either a numeric or number as a string with suffix input.}
}
\value{
\code{chr}
}
\description{
What is the human-readable suffix for a number
}
\examples{
num_suf(30000)
}
\seealso{
num2str
}
