% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{object_ext}
\alias{object_ext}
\title{Provide the appropriate file extension for a given object}
\usage{
object_ext(object)
}
\arguments{
\item{object}{to determine the appropriate function for writing to disk}
}
\value{
\code{(character)}
}
\description{
Provide the appropriate file extension for a given object
}
\examples{
object_ext(data.frame())
object_ext("asdf")
object_ext(structure("a", class = "ggplot"))
}
