% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{object_fn}
\alias{object_fn}
\title{Return the appropriate function for writing the supplied object to disk}
\usage{
object_fn(x, filepath)
}
\arguments{
\item{x}{\code{(object)}}
}
\value{
\code{(function)} See details for which function
}
\description{
Return the appropriate function for writing the supplied object to disk
}
\details{
\itemize{
\item{\code{data.frame/matrix}}{ Uses \link[arrow]{write_feather} if \code{arrow} installed. Otherwise uses \link[readr]{write_csv} if \code{readr} installed. Otherwise uses \link[base]{write.csv}}
\item{\code{ggplot}}{ \link[ggplot2]{ggsave}}
\item{\code{anything else}}{ \link[base]{saveRDS}}
}
}
\examples{
object_fn(1:15)
object_fn(data.frame())
object_fn(matrix())
object_fn(structure("x", class = "ggplot"))
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{load_obj}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
