% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{object_write}
\alias{object_write}
\title{Provide the appropriate file read/write function}
\usage{
object_write(x, write_path, mkpath = FALSE)
}
\arguments{
\item{x}{\code{(object)} to write to disk}

\item{write_path}{\code{(character)} path to write the object to}

\item{mkpath}{\code{(logical)} Whether to create the directory for \code{write_path} if it doesn't exist. \strong{Default} \code{FALSE}}
}
\value{
Success message if file is written
}
\description{
Write an object to disk
}
