% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{odds}
\alias{odds}
\title{Get odd numbers}
\usage{
odds(x)
}
\arguments{
\item{x}{\code{int}}
}
\description{
Get odd numbers
}
\examples{
odds(1:10)
}
\seealso{
Other math: 
\code{\link{comparison_inverse}()},
\code{\link{comparison_inverse_key}},
\code{\link{comparison_key}},
\code{\link{evens}()},
\code{\link{interpolate}()},
\code{\link{str_comparison}()}
}
\concept{math}
