% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_opts.R
\docType{data}
\name{opts}
\alias{opts}
\title{Check option value.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
opts\$your_option_name()
}
\arguments{
\item{default}{if the specified option is not set in the options list, this value is returned. This facilitates retrieving an option and checking whether it is set and setting it separately if not.}
}
\value{
\code{lgl}
}
\description{
This is a list that will populate dynamically with the options in the project local \emph{.Rprofile} allowing them to be read by calling the method. This population of methods happens at the beginning of a session. It can be accessed with \code{UU::opts}. If you wish to check option values in non-interactive sessions, see the \link[UU]{write_opts} function which will write a file with an \code{opts} object based on the current state of \emph{.Rprofile} that can be used during non-interactive (deployed apps) sessions.
}
\seealso{
write_opts
Code is dynamically generated when package is loaded. These will not auto generate properly in a deployed shiny app. See \link{write_opts} for a non-interactive alternative.

Other options: 
\code{\link{if_debug}()},
\code{\link{toggle}},
\code{\link{write_opts}()}
}
\concept{options}
\keyword{datasets}
