% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{package_size}
\alias{package_size}
\title{Return the size of a package, or all packages in a folder}
\usage{
package_size(packages, path = .libPaths()[1])
}
\arguments{
\item{packages}{\code{chr} of package names}

\item{path}{\code{chr} with package folders in it such as \code{\link[base]{.libPaths}}}
}
\value{
\code{chr/tbl} depending on whether packages or path is provided
}
\description{
Return the size of a package, or all packages in a folder
}
\examples{
package_size()
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{load_obj}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
