% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{path_strip_shiny}
\alias{path_strip_shiny}
\title{Strip a file path to everything after resourcepath if shiny is running}
\usage{
path_strip_shiny(path, resourcepath = "www", leading_slash = FALSE)
}
\arguments{
\item{path}{\code{chr}}

\item{resourcepath}{\code{chr} A resource path specified in \emph{app_ui.R}}

\item{leading_slash}{\code{lgl} If leading slash should be included \code{TRUE} or removed \code{FALSE}.}
}
\value{
\code{chr} without stripped directories
}
\description{
Useful for linking to internal files, such as with image source attributes \verb{<img src="[path]">}
}
\examples{
path_strip_shiny("inst/app/www/img/image.jpg")
}
\seealso{
Other shiny: 
\code{\link{as_js}()},
\code{\link{glue_js}()},
\code{\link{nm_to_id}()},
\code{\link{read_js}()},
\code{\link{shiny_error_recover}()},
\code{\link{strip_html}()}
}
\concept{shiny}
