% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_package_dev.R
\name{pkg_chr_split_comma}
\alias{pkg_chr_split_comma}
\title{Split a list of packages separated by commas}
\usage{
pkg_chr_split_comma(x)
}
\arguments{
\item{x}{\code{chr} of packages, often the output from \code{renv}}
}
\value{
\code{chr} vector with each package as a character
}
\description{
Split a list of packages separated by commas
}
\examples{
string_split_comma(c("DT, USAboundaries, USAboundariesData, cicerone, config",
"ggplot2, golem, googlesheets4, htmlwidgets, leaflet",
"paletteer, plotly, rmapshaper, scales, sf, shiny, shinyBS",
"shinyWidgets, shinyauthr, shinycssloaders, shinydashboard",
"shinyjs, shinyvalidate, tidyr") )
}
\seealso{
Other package dev: 
\code{\link{get_package_fns}()},
\code{\link{is_package}()},
\code{\link{is_package_dev}()},
\code{\link{need_pkg}()}
}
\concept{package dev}
