% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{possible_dupes}
\alias{possible_dupes}
\title{Detect possible duplicates after a join}
\usage{
possible_dupes(before, after, halt_fn = rlang::warn)
}
\arguments{
\item{before}{\code{(data.frame)} from before the join}

\item{after}{\code{(data.frame)} after the join}

\item{halt_fn}{\code{(function)} to notify, default \link[rlang]{warn}.}
}
\description{
Detect possible duplicates after a join
}
\examples{
a = data.frame(a = c(1, 2, 3, 4, 3, 5), b = 1:6)
b = data.frame(a = c(1, 2, 3, 4, 5), c = letters[1:5])
after <- dplyr::left_join(a, b)
possible_dupes(b, after, halt_fn = message)
}
\seealso{
\code{\link[rlang:abort]{rlang::abort()}} \code{\link[base:message]{base::message()}}
}
