% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{read_js}
\alias{read_js}
\title{Read Javascript file}
\usage{
read_js(filename)
}
\arguments{
\item{filename}{\code{chr}}
}
\value{
\code{chr}
}
\description{
Read Javascript file
}
\seealso{
Other shiny: 
\code{\link{as_js}()},
\code{\link{glue_js}()},
\code{\link{nm_to_id}()},
\code{\link{path_strip_shiny}()},
\code{\link{path_strip_to}()},
\code{\link{shiny_error_recover}()},
\code{\link{strip_html}()}
}
\concept{shiny}
