% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{\%|\%}
\alias{round_any}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{plyr}{\code{\link[plyr]{round_any}}}

  \item{rlang}{\code{\link[rlang:op-na-default]{\%|\%}}, \code{\link[rlang:op-null-default]{\%||\%}}}
}}

