% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{regex_op}
\alias{regex_op}
\title{Create a compound regex grouped statement}
\usage{
regex_op(x, type = "|", prefix = "", suffix = "")
}
\arguments{
\item{x}{\code{(character)} regex strings}

\item{type}{\code{(character)} \code{|}, \code{&}, \code{before}, \code{not_before} supported.}

\item{pre}{\code{(character)} regex tokens to precede each string group. IE \verb{(?:[pre]x)} w/out the braces}

\item{suf}{\code{(character)} regex tokens to follow each string group. IE \verb{(?:x[suf])} w/out the braces}
}
\value{
\code{(character)} grouped regex statement
}
\description{
Create a compound regex grouped statement
}
