% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{regex_or}
\alias{regex_or}
\title{Create a compound regex grouped OR statement}
\usage{
regex_or(x, pre = "", suf = "")
}
\arguments{
\item{x}{\code{(character)} regex strings}

\item{pre}{\code{(character)} regex tokens to precede each string group. IE \verb{(?:[pre]x)} w/out the braces}

\item{suf}{\code{(character)} regex tokens to follow each string group. IE \verb{(?:x[suf])} w/out the braces}
}
\value{
\code{(character)} grouped regex OR statement
}
\description{
Create a compound regex grouped OR statement
}
