% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert r,g,b,a values as string or numeric to hex}
\usage{
rgb2hex(red, green, blue, alpha = FALSE)
}
\arguments{
\item{red}{\code{chr/num} Either a CSS \code{rgb()} or \code{rgba()} declaration as a string, or the red value as numeric.}

\item{alpha}{\code{lgl/num} Whether to include the alpha value (an 8 digit hex) in the output or not, or the alpha value to apply, If set to TRUE, and alpha is not set, an alpha value of 1 will be appended. If numeric, a value between 0 & 1 inclusive to set as the alpha value.}
}
\value{
\code{chr} The hex value
}
\description{
Convert r,g,b,a values as string or numeric to hex
}
\examples{
rgb2hex("rgba(18,180,211,1)", alpha = TRUE)
rgb2hex("rgba(18,180,211,1)", alpha = FALSE)
}
\seealso{
rgb2hex_

Other color: 
\code{\link{color_cycle}()},
\code{\link{color_distance}()},
\code{\link{color_luminance}()},
\code{\link{color_match}()},
\code{\link{color_rgb_table}()},
\code{\link{color_separate}()},
\code{\link{color_text_by_luminance}()},
\code{\link{colors2css}()},
\code{\link{css_col2vec}()},
\code{\link{luminance_filter}()}
}
\concept{color}
