% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vectors.R
\name{rle_df}
\alias{rle_df}
\title{rle_df - create a run-length-encoding data.frame}
\usage{
rle_df(x)
}
\arguments{
\item{x}{\code{(vector)} An object for which to run an \code{rle}}
}
\value{
\item{(data.frame)}{ with length, values, start and end indices.}
}
\description{
Given an \code{\link[base]{rle}} this function will return a data.frame of starts, ends, and indexes thereof of the run lengths.
Credit: \url{https://stackoverflow.com/questions/43875716/find-start-and-end-positions-indices-of-runs-consecutive-values}
}
\examples{
rle_df(sample(c(TRUE,FALSE), replace = TRUE, 100))
}
\seealso{
Other vectors: 
\code{\link{len_unique}()},
\code{\link{names_values_switch}()},
\code{\link{rle_groups}()},
\code{\link{rle_seq}()},
\code{\link{sort_by_names}()},
\code{\link{true_names}()},
\code{\link{unify_vec_preserve_order}()},
\code{\link{unique_with_names}()},
\code{\link{vlookup}()},
\code{\link{vlookup_from_ref}()},
\code{\link{zchar_remove}()}
}
\concept{vectors}
