% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vectors.R
\name{rle_groups}
\alias{rle_groups}
\title{Create an RLE Grouping from a logical vector}
\usage{
rle_groups(x)
}
\arguments{
\item{x}{\code{lgl} vector}
}
\value{
\code{list}
}
\description{
Create an RLE Grouping from a logical vector
}
\examples{
rle_groups(sample(c(TRUE, FALSE), 20, replace = TRUE))
}
\seealso{
Other vectors: 
\code{\link{len_unique}()},
\code{\link{names_values_switch}()},
\code{\link{rle_df}()},
\code{\link{rle_seq}()},
\code{\link{sort_by_names}()},
\code{\link{true_names}()},
\code{\link{unify_vec_preserve_order}()},
\code{\link{unique_with_names}()},
\code{\link{vlookup_from_ref}()},
\code{\link{zchar_remove}()}
}
\concept{vectors}
