% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rounding.R
\name{round_to}
\alias{round_to}
\title{Find convenient limits of input vectors}
\usage{
round_to(
  ...,
  accuracy = NULL,
  fn = min,
  f = if (identical(fn, min)) {
     floor
 } else {
     ceiling
 },
  na.rm = TRUE
)
}
\arguments{
\item{...}{\code{num} vectors}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{fn}{\code{fun} \link[base]{min}/\link[base]{max}}

\item{f}{rounding function: \code{\link{floor}}, \code{\link{ceiling}} or
\code{\link{round}}}

\item{na.rm}{a logical indicating whether missing values should be
    removed.}
}
\value{
\code{num}
}
\description{
If accuracy is omitted, number will be rounded to the nearest order of magnitude IE 145, if \code{fn = min}, will round to 100
}
\examples{
round_to(runif(10, 5, 10), sample(1:10, 5))
round_to(runif(10, 5, 10), sample(1:10, 5), fn = max)
round_to(45)
round_to(145)
round_to(145, fn = max)
}
\seealso{
Other rounding: 
\code{\link{magnitude_order}()},
\code{\link{magnitude_triplet}()},
\code{\link{num2str}()},
\code{\link{num2str_vec}()},
\code{\link{num_chr_suffi}},
\code{\link{size}()},
\code{\link{unit_find}()},
\code{\link{unit_modify}()},
\code{\link{unit_modify_vec}()},
\code{\link{unit_string}()}
}
\concept{rounding}
