% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date_time.R
\name{season_factor}
\alias{season_factor}
\title{Season as factor/numeric}
\usage{
season_factor(x, label = FALSE, abbr = TRUE)
}
\arguments{
\item{x}{\code{chr/Date/POSIXt} Names of seasons as case insensitive character vector, if using abbreviations, they must be of consistent length. Otherwise, a \code{Date} or \code{POSIXt}.}
}
\value{
\code{factor} If \code{x} is provided, the season as a factor. If \code{x} is not provided, an ordered factor of the seasons.
}
\description{
Season as factor/numeric
}
\examples{
season_factor()
season_factor(c('Sp', 'Su'))
season_factor(c('Sp', 'Su'), label = TRUE)
season_factor(seq(lubridate::floor_date(Sys.time(), "year"), Sys.time(), by = "day"), label = TRUE, abbr = FALSE)
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{excel_date}()},
\code{\link{month_factor}()},
\code{\link{time_aggregates}},
\code{\link{time_difftimes}},
\code{\link{time_elapsed}()},
\code{\link{time_factor}()},
\code{\link{timespan}()},
\code{\link{week_factor}()}
}
\concept{time}
