% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{shiny_error_recover}
\alias{shiny_error_recover}
\title{Toggle \link[utils]{recover} on error when obtuse shiny errors are encountered}
\usage{
shiny_error_recover()
}
\description{
Toggle \link[utils]{recover} on error when obtuse shiny errors are encountered
}
