% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{smode}
\alias{smode}
\title{Statistical mode}
\usage{
smode(x)
}
\arguments{
\item{x}{\code{(vector)}}
}
\value{
\code{(vector)}
}
\description{
Return the most frequenctly occuring item in a dataset
}
\seealso{
Other statistics: 
\code{\link{max2}()}
}
\concept{statistics}
