% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{sort_by_names}
\alias{sort_by_names}
\title{Sort a vector or list by it's name (or self if no names)}
\usage{
sort_by_names(x, by_names = TRUE)
}
\arguments{
\item{x}{\code{obj} to sort}

\item{by_names}{\code{lgl} wether to sort by names}
}
\value{
\code{obj} sorted
}
\description{
Sort a vector or list by it's name (or self if no names)
}
\examples{
sort_by_names(c(b = "b", c = "a"))
sort_by_names(c(b = "b", c = "a"), by_names = FALSE)
}
