% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{start_cluster}
\alias{start_cluster}
\title{start_cluster}
\usage{
start_cluster(
  workers = future::availableCores()\%/\%2,
  timeout = 60 * 60 * 5,
  outfile = file.path(getwd(), "cl_out.log")
)
}
\arguments{
\item{workers}{\code{(numeric)} number of worker nodes. Defaults to 3/4 of the available.}

\item{outfile}{\code{(character/logical)} Path and name of the outfile or \code{FALSE} to disable an outfile.}
}
\value{
\code{(RichSOCKcluster)}
}
\description{
Creates a compute cluster
}
