% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{str_comparison}
\alias{str_comparison}
\title{Convert inequality statements between character, mathematic, symbol and function representations}
\usage{
str_comparison(x, outtype = "chr")
}
\arguments{
\item{x}{\code{chr} vector or inequality statements}

\item{outtype}{\code{chr} class of the outype
\itemize{
\item{\code{"chr"}} A character
\item{\code{"str"}} A character
\item{\code{"sym"}} A symbol/name
\item{\code{"name"}} A symbol/name
\item{\code{"fun"}} A function
}}
}
\value{
\code{chr/name/fun} depending on the requested \code{outtype}
}
\description{
Convert inequality statements between character, mathematic, symbol and function representations
}
\examples{
str_comparison(">")
str_comparison("less than or equal to", "fun")
str_comparison("greater than or equal to", "sym")
}
\seealso{
comparison_key
}
