% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\docType{data}
\name{time_aggregates}
\alias{time_aggregates}
\title{Timespans as character}
\format{
An object of class \code{list} of length 7.
}
\usage{
time_aggregates
}
\description{
Timespans as character
}
\keyword{datasets}
