% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\docType{data}
\name{time_difftimes}
\alias{time_difftimes}
\title{Timespans as durations}
\format{
An object of class \code{list} of length 7.
}
\usage{
time_difftimes
}
\description{
Timespans as durations
}
\keyword{datasets}
