% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\name{time_elapsed}
\alias{time_elapsed}
\title{Return a logical on an interval}
\usage{
time_elapsed(file = ".interval_timer.rds", interval = lubridate::weeks(1))
}
\arguments{
\item{file}{\code{chr} filename in which to store the interval time}

\item{interval}{\code{period/duration} Default 1 week}
}
\value{
\code{lgl}
}
\description{
Return a logical on an interval
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{excel_date}()},
\code{\link{month_factor}()},
\code{\link{season_factor}()},
\code{\link{time_aggregates}},
\code{\link{time_difftimes}},
\code{\link{time_factor}()},
\code{\link{timespan}()},
\code{\link{week_factor}()}
}
\concept{time}
