% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\name{time_factor}
\alias{time_factor}
\title{Turn timespans into an ordered factor}
\usage{
time_factor(x)
}
\arguments{
\item{x}{\code{chr} of timespans}
}
\value{
\code{factor}
}
\description{
Turn timespans into an ordered factor
}
\examples{
time_factor(c("year", "week"))
}
