% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_helpers.R
\docType{data}
\name{toggle}
\alias{toggle}
\title{Toggle an option}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
toggle
}
\arguments{
\item{opt}{\code{chr} option name to toggle. See \code{.virga_opts} for standard Virga options}

\item{set}{\code{obj} value to set the option to (takes precedence over toggle)}
}
\value{
\code{msg} to inform user of what the option is set to
}
\description{
Meant for interactive use only.
}
\keyword{datasets}
