% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_condition_signaling.R
\name{trace_back_json}
\alias{trace_back_json}
\title{Writes a trace back as a json for error logging
Useful for remote error logging on deployed shiny apps, such as via Sentry}
\usage{
trace_back_json(
  e = NULL,
  file = glue::glue("{lubridate::format_ISO8601(Sys.time())}.json"),
  tb = rlang::trace_back(bottom = 1)
)
}
\arguments{
\item{e}{\code{error} Error condition object, optional}

\item{file}{\code{chr} path to file to be written.}

\item{tb}{\code{rlang_trace} Trace back}
}
\value{
\code{None} called for side effect of writing to file
}
\description{
Writes a trace back as a json for error logging
Useful for remote error logging on deployed shiny apps, such as via Sentry
}
\examples{
trace_back_json()
}
\seealso{
Other condition signaling: 
\code{\link{gbort}()},
\code{\link{ginfo}()},
\code{\link{gmsg}()},
\code{\link{gwarn}()}
}
\concept{condition signaling}
