% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vectors.R
\name{true_names}
\alias{true_names}
\title{Return the names of all TRUE items in a logical vector}
\usage{
true_names(x)
}
\arguments{
\item{x}{\code{lgl} with names}
}
\value{
\code{chr} of the names of true values
}
\description{
Return the names of all TRUE items in a logical vector
}
\examples{
true_names(c(a = TRUE, b = FALSE))
true_names(c(a = FALSE, b = FALSE))
}
\seealso{
Other vectors: 
\code{\link{len_unique}()},
\code{\link{names_values_switch}()},
\code{\link{rle_df}()},
\code{\link{rle_groups}()},
\code{\link{rle_seq}()},
\code{\link{sort_by_names}()},
\code{\link{unify_vec_preserve_order}()},
\code{\link{unique_with_names}()},
\code{\link{vlookup}()},
\code{\link{vlookup_from_ref}()},
\code{\link{zchar_remove}()}
}
\concept{vectors}
