% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unique_with_names}
\alias{unique_with_names}
\title{Unique a vector, preserving the names of the first original entries}
\usage{
unique_with_names(x)
}
\arguments{
\item{x}{\code{vec} with names}
}
\value{
\code{vec} of the same type
}
\description{
Unique a vector, preserving the names of the first original entries
}
\examples{
unique_with_names(c(N = "n", b = "b", A = "n"))
}
