% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rounding.R
\docType{data}
\name{unit_conversion}
\alias{unit_conversion}
\title{Abbreviations of numeric magnitude for various units}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 6 columns.
}
\usage{
unit_conversion
}
\description{
Abbreviations of numeric magnitude for various units
}
\keyword{datasets}
