% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unit_find}
\alias{unit_find}
\title{Find the row corresponding to a value in \code{unit_conversion}}
\usage{
unit_find(x)
}
\arguments{
\item{x}{\code{chr/num} vector to find}
}
\value{
\code{tbl}
}
\description{
Find the row corresponding to a value in \code{unit_conversion}
}
\examples{
unit_find("B")
}
\seealso{
Other rounding: 
\code{\link{magnitude_order}()},
\code{\link{magnitude_triplet}()},
\code{\link{num2str_vec}()},
\code{\link{num2str}()},
\code{\link{num_chr_suffi}},
\code{\link{round_to}()},
\code{\link{size}()},
\code{\link{unit_modify_vec}()},
\code{\link{unit_modify}()},
\code{\link{unit_string}()}
}
\concept{rounding}
