% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unit_modify}
\alias{unit_modify}
\title{Modify unit abbreviations}
\usage{
unit_modify(x, unit, outtype)
}
\arguments{
\item{unit}{\code{chr} Type of units, supported values are: \code{AF, F, $, W, MW}}

\item{outtype}{\code{chr} The type of output to be added. Current possibilities are: \code{begin, end, abbrev, unit_eng}}

\item{magnitude}{\code{num} The order of magnitude for the unit. The highest triplet will be used. See \code{magnitude_triplet}}
}
\value{
\code{chr}
}
\description{
Modify unit abbreviations
}
\examples{
unit_modify("AF", 10^7, "abbrev")
}
