% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unit_modify_vec}
\alias{unit_modify_vec}
\title{Modify unit abbreviation, vectorized version}
\usage{
unit_modify_vec(x, unit, outtype)
}
\arguments{
\item{unit}{\code{chr} Type of units, supported values are: \code{AF, F, $, W, MW}}

\item{outtype}{\code{chr} The type of output to be added. Current possibilities are: \code{begin, end, abbrev, unit_eng}}
}
\value{
\code{chr} updated units
}
\description{
Modify unit abbreviation, vectorized version
}
\examples{
unit_modify(10^7, "AF", "abbrev")
}
\seealso{
unit_modify
}
