% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unit_select}
\alias{unit_select}
\title{Select unit abbreviations}
\usage{
unit_select(x, unit, outtype)
}
\arguments{
\item{unit}{\code{chr} Type of units, supported values are: \code{AF, F, $, W, MW}}

\item{outtype}{\code{chr} The type of output requested. Current possibilities are: \code{begin, end, abbrev, unit_eng}}

\item{magnitude}{\code{num} The order of magnitude for the unit. The highest triplet will be used. See \code{magnitude_triplet}}
}
\value{
\code{chr}
}
\description{
Select unit abbreviations
}
\examples{
unit_select("AF", 10^7, "abbrev")
}
