% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unit_shorthand}
\alias{unit_shorthand}
\title{Easily translate long form unit names to shorthand}
\usage{
unit_shorthand(x, units = unit_trans)
}
\arguments{
\item{x}{\code{chr/tbl} Character or data.frame with long form names. Currently supports Acre-feet = AF,Months = mths,Percent = \%,Years = yrs,Number = #,Cubic feet per second = Ft^3/s,Million Acre-Feet = MAF,Megawatt-hours = MWh,Megawatts = MW,Gigawatt-hours = GWh,Gigawatt = GW,Feet = Ft}

\item{units}{\code{chr} Unit translation vector with shorthand as the vectur and long form names as the names. Uses \code{unit_trans} as default, append additional for specific translations not yet represent in \code{unit_trans} and open a PR or issue to add the translations if inclined.}
}
\value{
\code{chr/tbl} with same class as \code{x}
}
\description{
Useful for condensed displays like axis titles
}
\examples{
unit_shorthand(tibble::tibble("Max Gigawatt-hours" = 5, "Really big number" = 10^6))
}
