% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unit_string}
\alias{unit_string}
\alias{unit_string<-}
\title{Extract the units from a string}
\usage{
unit_string(x)

unit_string(x) <- value
}
\arguments{
\item{x}{\code{chr} String(s) to extract units from/assign units to}
}
\value{
\code{chr}
}
\description{
It is assumed that units are encased in parentheses at the end of the string
}
\examples{
unit_string("Elevation (F)")
}
