% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unload_namespaces}
\alias{unload_namespaces}
\title{Unload namespaces prior to package install}
\usage{
unload_namespaces(ns)
}
\arguments{
\item{ns}{\code{(chr)} namespaces to unload}
}
\description{
Unload namespaces prior to package install
}
