% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_imports.R
\name{use_reimport}
\alias{use_reimport}
\title{Add a function to reimports}
\usage{
use_reimport(pkg, fun)
}
\arguments{
\item{pkg}{\code{chr} package name}

\item{fun}{\code{chr} function name}
}
\value{
\code{msg}
}
\description{
Add a function to reimports
}
