% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{use_UU_reimports}
\alias{use_UU_reimports}
\title{Write \emph{R/aaa_reimports.R} file}
\usage{
use_UU_reimports(file = "R/aaa_reimports.R")
}
\arguments{
\item{file}{\code{chr} path to file to write}
}
\description{
Write \emph{R/aaa_reimports.R} file
}
