% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{vlookup}
\alias{vlookup}
\title{Simple lookup of values}
\usage{
vlookup(x, lookup)
}
\arguments{
\item{x}{\code{any} Values to lookup}

\item{lookup}{\code{named any} names will be used as replacement}
}
\value{
\code{any}
}
\description{
Simple lookup of values
}
\examples{
lookup <- rlang::set_names(1:5, letters[1:5])
vlookup(sample(1:5, 5), lookup)
}
