% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{vlookup_from_ref}
\alias{vlookup_from_ref}
\title{Vlookup replace using a lookup column and reference table}
\usage{
vlookup_from_ref(base, lookup_col, lookup_ref, value_col)
}
\arguments{
\item{base}{\code{vector} of starting values. Replacements will be made in this vector before it is returned.}

\item{lookup_col}{\code{vector} of values with same length as \code{base} that will be matched to \code{lookup_ref} to deteremine the replacement indices}

\item{lookup_ref}{\code{vector} of reference values, which \code{lookup_col} will be matched to in order to determine replacement values.}

\item{value_col}{\code{vector} of replacement values with same length as \code{lookup_ref}}
}
\value{
\code{vector}
}
\description{
Vlookup replace using a lookup column and reference table
}
\examples{
ref <- tibble::tibble(lookup = letters[1:5], value = 1:5)
original <- tibble::tibble(lookup = letters[1:20], base = runif(20, min = 6, max = 20))
dplyr::mutate(original, base = vlookup_from_ref(base, lookup, ref$lookup, ref$value))
}
