% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date_time.R
\name{week_bins_per_year}
\alias{week_bins_per_year}
\title{Create a tbl with each day of the year, it's week number, and the interval containing that week.}
\usage{
week_bins_per_year(year, inclusive = TRUE)
}
\arguments{
\item{year}{\code{int} Year number}

\item{inclusive}{\code{lgl} Whether to include any days from the previous and next years in the first and final weeks of the year respectively.}
}
\value{
\code{tbl} With features:
\itemize{
\item{\code{day}}{ The Date of each day}
\item{\code{wn}}{ week number. See \code{\link[lubridate]{isoweek}} for details.}
\item{\code{interval}}{ The \code{\link[lubridate]{interval}} for the week.}
}
}
\description{
Create a tbl with each day of the year, it's week number, and the interval containing that week.
}
