% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\name{week_factor}
\alias{week_factor}
\title{Get numeric day of the week}
\usage{
week_factor(x)
}
\arguments{
\item{x}{\code{chr} Days of the week. Abbreviations are fine, case insensitive}
}
\value{
\code{factor} If \code{x} is provided, the day of the week as a factor. If \code{x} is not provided, an ordered factor with Monday as 1.
}
\description{
Get numeric day of the week
}
\examples{
week_factor()
week_factor('Tu')
}
