% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{write_lines}
\alias{write_lines}
\title{Write lines at a specific location in a file}
\usage{
write_lines(file, ..., after = "end")
}
\arguments{
\item{file}{\code{chr} path to file}

\item{...}{\code{chr} lines to write}

\item{after}{\code{num/chr} either a line number of "end" to write it at the end}
}
\value{
the resulting file
}
\description{
Write lines at a specific location in a file
}
