% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_opts.R
\name{write_opts}
\alias{write_opts}
\title{Write all the option checking functions to a file}
\usage{
write_opts(file = "R/utils_opts.R")
}
\arguments{
\item{file}{\code{chr} Default \emph{R/opts.R}}
}
\value{
\code{msg}
}
\description{
Write all the option checking functions to a file
}
\seealso{
Other options: 
\code{\link{if_debug}()},
\code{\link{opts}},
\code{\link{toggle}}
}
\concept{options}
