% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{write_to_rprofile}
\alias{write_to_rprofile}
\title{Write expressions to the \emph{.Rprofile}}
\usage{
write_to_rprofile(..., scope = c("user", "project")[1])
}
\arguments{
\item{...}{\code{exprs}}

\item{scope}{\code{chr} which \emph{.Rprofile} to write to}
}
\value{
\code{msg}
}
\description{
Write expressions to the \emph{.Rprofile}
}
\seealso{
Other project setup: 
\code{\link{creds_to_renviron}()},
\code{\link{fun_docs_table}()},
\code{\link{ignore_files}()},
\code{\link{install_remote}()},
\code{\link{key_pairs_duplicated}()},
\code{\link{key_pairs_text}()},
\code{\link{use_UU_reimports}()},
\code{\link{use_reimport}()}
}
\concept{project setup}
