% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditionals.R
\name{zchar}
\alias{zchar}
\title{Is zero-length character?}
\usage{
zchar(...)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{lgl}
}
\description{
Is zero-length character?
}
\examples{
zchar("")
zchar(" ")
}
\seealso{
Other conditionals: 
\code{\link{is_error}()},
\code{\link{is_legit}()},
\code{\link{is_project}()},
\code{\link{larger}()},
\code{\link{most}()},
\code{\link{same}()}
}
\concept{conditionals}
