% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vectors.R
\name{zchar_remove}
\alias{zchar_remove}
\title{Remove zero length strings (or string with all spaces)}
\usage{
zchar_remove(x)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{chr}
}
\description{
Remove zero length strings (or string with all spaces)
}
\examples{
zchar_remove(c("", "  ", "a"))
}
\seealso{
Other vectors: 
\code{\link{len_unique}()},
\code{\link{names_values_switch}()},
\code{\link{rle_df}()},
\code{\link{rle_groups}()},
\code{\link{rle_seq}()},
\code{\link{sort_by_names}()},
\code{\link{true_names}()},
\code{\link{unify_vec_preserve_order}()},
\code{\link{unique_with_names}()},
\code{\link{vlookup_from_ref}()}
}
\concept{vectors}
