% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{zero-length-infix}
\alias{zero-length-infix}
\alias{\%|0|\%}
\title{Replace a 0 length value}
\usage{
lhs \%|0|\% rhs
}
\arguments{
\item{lhs}{\code{(expression)} to try}

\item{rhs}{\code{(expression)} to replace if expression fails}
}
\value{
results from lhs if length > 1 otherwise rhs
}
\description{
If the lhs is length 0, replace with rhs
}
