% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{as_js}
\alias{as_js}
\title{Preserve a string as JS/HTML (prevent translation of characters)}
\usage{
as_js(x)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{chr, HTML, JS_EVAL}
}
\description{
Preserve a string as JS/HTML (prevent translation of characters)
}
\seealso{
\code{\link[DT:DT-imports]{DT::JS()}}

Other shiny: 
\code{\link{glue_js}()},
\code{\link{nm_to_id}()},
\code{\link{path_strip_shiny}()},
\code{\link{path_strip_to}()},
\code{\link{read_js}()},
\code{\link{shiny_error_recover}()},
\code{\link{strip_html}()}
}
\concept{shiny}
