% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environments_namespaces.R
\name{assign_in_ns}
\alias{assign_in_ns}
\title{Assign a variable into a namespace}
\usage{
assign_in_ns(
  x,
  nm = rlang::expr_deparse(rlang::enexpr(x)),
  ns_env = rlang::ns_env(pkg_name()),
  x_is_obj_name = TRUE
)
}
\arguments{
\item{x}{\code{object/chr} either the object itself or the name of the object to assign}

\item{nm}{\code{chr} name for object in the namespace}

\item{ns_env}{\code{env} of the namespace}

\item{x_is_obj_name}{\code{lgl} If x is a character, is it the name of the object? Set to \code{FALSE} if x is a character object that you wish to assign \emph{and not the name of the object}.}
}
\value{
\code{x} the object
}
\description{
Unlocks and relocks namespaces and bindings as needed
}
\seealso{
Other namespaces: 
\code{\link{assign_global}()},
\code{\link{get_from_ns}()},
\code{\link{get_global}()},
\code{\link{pkg_ns}()},
\code{\link{unload_namespaces}()}
}
\concept{namespaces}
