% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_helpers.R
\name{assign_in_ns}
\alias{assign_in_ns}
\title{Assign a variable into a namespace}
\usage{
assign_in_ns(
  x,
  nm = rlang::expr_deparse(rlang::enexpr(x)),
  ns_env = rlang::ns_env(pkg_name())
)
}
\arguments{
\item{x}{\code{object} to assign}

\item{nm}{\code{chr} name for object in the namespace}

\item{ns_env}{\code{env} of the namespace}
}
\value{
\code{none}
}
\description{
Unlocks and relocks namespaces and bindings as needed
}
