% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{col_types}
\alias{col_types}
\title{Converts input to a specified type output}
\usage{
col_types(x, outtype = c("chr", "hud", "fun", "typ", "col")[1])
}
\arguments{
\item{x}{\code{(vector/function)} One of:
\itemize{
\item{column}{ \code{(any)}}
\item{a type specification from HUD}{ \code{(character)}}
\item{a readr \verb{parse_*} function (See \link[readr]{parse_logical})}{ \code{(function)}}
\item{a readr type specification (See \link[readr]{cols})}{ \code{(character)}}
}}

\item{outtype}{\code{(character)} One of:
\itemize{
\item{\code{"chr"}}{ Returns the class as a readr abbreviation (See \link[readr]{cols})}
\item{\code{"hud"}}{ \code{(character)} a type specification from HUD}
\item{\code{"fun"}}{a readr \verb{parse_*} function (See \link[readr]{parse_logical})}{ \code{(function)}}
\item{\code{"typ"}}{ \code{(character)} The R data class}
\item{\code{"col"}}{ \code{(character)} The \code{\link[readr]{collector}}}
}}
}
\value{
See outtype
}
\description{
Given various inputs, provide a col_type specification in the format indicated by \code{outtype}
}
\seealso{
Other file IO: 
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
