% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_cycle}
\alias{color_cycle}
\title{Makes a cyclic color palette of a specified length using the specified transformation each cycle}
\usage{
color_cycle(colors, n, transform_fn = colorspace::lighten, ...)
}
\arguments{
\item{colors}{\code{chr} vector of colors in hex format}

\item{n}{\code{num} length of color vector needed}

\item{transform_fn}{\code{fun} See \link\link{colorspace} for examples}

\item{...}{\code{args} passed on to \code{transform_fn}}
}
\value{
\code{chr} vector of \code{n} length
}
\description{
Makes a cyclic color palette of a specified length using the specified transformation each cycle
}
\examples{
color_cycle(c("rgba(18,180,211,1)", "rgba(2,120,170,1)", "rgba(0,57,73,1)"), n = 9, amount = .2)
}
