% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_distance}
\alias{color_distance}
\title{Compute color distance}
\usage{
color_distance(x, y)
}
\arguments{
\item{x}{\code{num/chr} A CSS hex or rgb/rgba color, or a numeric vector of r,g,b values}

\item{y}{\code{num/chr} A CSS hex or rgb/rgba color, or a numeric vector of r,g,b values}
}
\value{
\code{num} A representation of the distance for comparison with other distances
}
\description{
Compute color distance
}
\examples{
color_distance("rgba(111,96,140,1)", "#12B4D3")
}
