% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_luminance}
\alias{color_luminance}
\title{Find the luminance of a particular color, scaled 0-1}
\usage{
color_luminance(red, green, blue, alpha, ...)
}
\arguments{
\item{red, blue, green, alpha}{numeric vectors with values in \eqn{[0, M]}
    where \eqn{M} is \code{maxColorValue}.  When this is \code{255}, the
    \code{red}, \code{blue}, \code{green}, and \code{alpha}
    values are coerced to integers in \code{0:255} and the result is
    computed most efficiently.}

\item{...}{additional arguments
See \code{color_text_by_luminance} for creating contrasting text color to a background.}
}
\value{
\code{num}
}
\description{
Find the luminance of a particular color, scaled 0-1
}
\examples{
rlang::exec(color_luminance, !!!css_col2vec('white'))
}
\seealso{
\code{\link[=color_text_by_luminance]{color_text_by_luminance()}}

Other color: 
\code{\link{color_cycle}()},
\code{\link{color_distance}()},
\code{\link{color_match}()},
\code{\link{color_rgb_table}()},
\code{\link{color_separate}()},
\code{\link{color_text_by_luminance}()},
\code{\link{colors2css}()},
\code{\link{css_col2vec}()},
\code{\link{luminance_filter}()},
\code{\link{rgb2hex}()}
}
\concept{color}
