% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_match}
\alias{color_match}
\title{Match colors by visual distance}
\usage{
color_match(x, y, with_replacement = TRUE)
}
\arguments{
\item{x}{\code{chr} Named CSS representations of colors that will be matched}

\item{y}{\code{chr} Named CSS representations of colors that will be selected from as matches}

\item{with_replacement}{\code{lgl} Whether colors in x should have unique matches in y, or if y can be matched with replacement. Default TRUE to match with the closest color in y, regardless if it's already been matched with a previous color in x.}
}
\value{
\code{tbl} Of matches for all of \code{x} with the associated distance
}
\description{
Helpful for pairing colors across light/dark palettes
}
\examples{
a <- list(
  light = list(
    aquamarine = "rgb(137,210,215)",
    teal = "rgb(79,194,198)",
    turquoise = "rgb(3,119,132)",
    evergreen = "rgb(0,84,92)",
    brown = "rgb(72,36,18)",
    lightbrown = "rgb(132,96,78)",
    orange = "rgb(250,173,25)",
    darkorange = "rgb(198,143,44)",
    aliceblue = "rgb(180,243,249)",
    gainsboro = "rgb(223,254,255)"
  ),
  dark = list(
    lightblue = "rgba(18,180,211,1)",
    darkblue = "rgba(2,120,170,1)",
    navyblue = "rgba(0,57,73,1)",
    brown = "rgba(72,36,18,1)",
    lightbrown = "rgb(132,96,78)",
    lighterbrown = "rgba(181,141,122, 1)",
    purple = "rgba(111,96,140,1)",
    lightpurple = "rgba(165,150,194,1)",
    darkcyan = "rgba(0,166,212,1)",
    darkturquoise = "rgba(9,119,168,1)"
  )
)
color_match(a$light, a$dark)
}
\seealso{
Other color: 
\code{\link{color_cycle}()},
\code{\link{color_distance}()},
\code{\link{color_luminance}()},
\code{\link{color_rgb_table}()},
\code{\link{color_separate}()},
\code{\link{color_text_by_luminance}()},
\code{\link{colors2css}()},
\code{\link{css_col2vec}()},
\code{\link{luminance_filter}()},
\code{\link{rgb2hex}()}
}
\concept{color}
