% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_rgb_table}
\alias{color_rgb_table}
\title{Convert vector of colors to named tbl}
\usage{
color_rgb_table(colors)
}
\arguments{
\item{colors}{\code{chr} vector of css color representations}
}
\value{
\code{tbl} Columns \code{red}, \code{green}, \code{blue}, \code{alpha} and each color is a row
}
\description{
Convert vector of colors to named tbl
}
\examples{
color_rgb_table(c("red", "green", "#45fa32"))
}
