% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{color_separate}
\alias{color_separate}
\title{Separate a vector of colors based on their distance}
\usage{
color_separate(x)
}
\arguments{
\item{x}{\code{chr} Vector of hex or rgb/rgba color values}
}
\value{
\code{chr} The original vector, sorted for contrast
}
\description{
Separate a vector of colors based on their distance
}
\examples{
color_separate(c("rgba(18,180,211,1)", "rgba(2,120,170,1)", "rgba(0,57,73,1)",
"rgba(72,36,18,1)", "rgba(111,96,140,1)", "rgba(0,166,212,1)",
"rgba(9,119,168,1)"))
}
