% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_css_color.R
\name{colors2css}
\alias{colors2css}
\title{Convert a list of colors to SCSS/Sass variables or classes}
\usage{
colors2css(colors, file = "", sass_vars = TRUE)
}
\arguments{
\item{file}{A \link[base]{connection}, or a character string naming the file
    to print to.  If \code{""} (the default), \code{cat} prints to the
    standard output connection, the console unless redirected by
    \code{\link[base]{sink}}.
    If it is \code{"|cmd"}, the output is piped to the command given
    by \file{cmd}, by opening a pipe connection.
  }

\item{color_theme}{\code{2-layer list} with dark/light themes}
}
\value{
\code{msg} Sass/SCSS formatted variables or classes
}
\description{
Convert a list of colors to SCSS/Sass variables or classes
}
\examples{
colors2css(c(a = "white", b = "green"))
}
\seealso{
Other color: 
\code{\link{color_cycle}()},
\code{\link{color_distance}()},
\code{\link{color_luminance}()},
\code{\link{color_match}()},
\code{\link{color_rgb_table}()},
\code{\link{color_separate}()},
\code{\link{color_text_by_luminance}()},
\code{\link{css_col2vec}()},
\code{\link{luminance_filter}()},
\code{\link{rgb2hex}()}
}
\concept{color}
