% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{common_names}
\alias{common_names}
\title{Find the common names between two objects}
\usage{
common_names(...)
}
\arguments{
\item{...}{\code{(objects)}}
}
\value{
\code{(character)} of the common names
}
\description{
Given named objects, find the names in common
}
\examples{
common_names(rlang::set_names(letters), rlang::set_names(letters[4:10]))
}
