% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{comparison_inverse}
\alias{comparison_inverse}
\title{Convert a math comparator to it's inverse}
\usage{
comparison_inverse(x)
}
\arguments{
\item{x}{\code{chr/fun} Math comparator, one of <, >, <=, >=}
}
\value{
\code{chr/fun} the inverse comparator as the same class as \code{x}
}
\description{
Convert a math comparator to it's inverse
}
\examples{
comparison_inverse(">=")
}
\seealso{
Other math: 
\code{\link{comparison_inverse_key}},
\code{\link{comparison_key}},
\code{\link{evens}()},
\code{\link{interpolate}()},
\code{\link{odds}()},
\code{\link{str_comparison}()}
}
\concept{math}
