% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\docType{data}
\name{comparison_inverse_key}
\alias{comparison_inverse_key}
\title{Math comparison comparator inverse key}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
comparison_inverse_key
}
\description{
Math comparison comparator inverse key
}
\keyword{datasets}
