% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\docType{data}
\name{comparison_key}
\alias{comparison_key}
\title{Math comparison comparator to plain english key}
\format{
An object of class \code{character} of length 7.
}
\usage{
comparison_key
}
\description{
Math comparison comparator to plain english key
}
\keyword{datasets}
