% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{concat_rows}
\alias{concat_rows}
\title{Concatenate row values in a poorly scraped table}
\usage{
concat_rows(.data, col_to_check = 1)
}
\arguments{
\item{.data}{\code{tbl} Of data with empty rows}

\item{col_to_check}{\code{num} The column with rows populated into which subsequent rows will be collapsed.}
}
\value{
\code{tbl}
}
\description{
Concatenate row values in a poorly scraped table
}
\examples{
test_tbl <- tibble::tribble(
~Facility, ~State,           ~City,    ~Production,             ~Closure.Date.Facility.Owners,         ~Water.Use, ~Water.Rights.Amount..Type, ~Water.Rights.Owners,
"Coronado Generating Station",   "AZ",    "St . Johns",       "762 MW", "Units 1 and 2: Salt River Project (SRP)",      "5,200 AF/YR",            "44,000 AF/YR,",                #'"SRP",
NA,     NA,              NA,             NA,                             "2032 (100\%)",                 NA,                  "GW only",                   NA,
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
"Springerville Generating",   "AZ", "Springerville",     "1,625 MW",      "Unit 1: 2027 Tucson Electric (TEP)",     "10,400 AF/YR",            "77,000 AF/YR,",          "SRP, TEP#',",
"Station",     NA,              NA,             NA,                   "(50\%, PC: UNS Energy)",                 NA,                  "GW only",      "Tri-State G&T",
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                            "Unit 2: 2032",                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                         "Tri-State (25\%)",                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                          "Units 3 and 4:",                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                           "TBD SRP (25\%)",                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
"Cholla Power Plant",   "AZ",   "Joseph City",       "767 MW",   "Units 1 and 2: Arizona Public Service",      "9,200 AF/YR",             "9,550 AF/YR,",                "APS",
NA,     NA,              NA,             NA,                      "2025 Company (APS)",                 NA,                  "GW only",                   NA,
NA,     NA,              NA,             NA,   "(100\%, PC: Pinnacle West)Unit 3: 2020",                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
NA,     NA,              NA,             NA,                                        NA,                 NA,                         NA,                   NA,
"Kayenta Mine",   "AZ",       "Kayenta", "6 .5 million",           "Unknown Peabody Energy (100\%)",          "Unknown",            "11,200 AF/YR,",            "Peabody",
NA,     NA,              NA,   "short tons",                                        NA, "(est . less than",                  "GW only",                   NA,
NA,     NA,              NA,             NA,                                        NA,       "100 AF/YR)",                         NA,                   NA
)
(out <- concat_rows(test_tbl, State))
}
