% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_setup.R
\name{creds_to_renviron}
\alias{creds_to_renviron}
\title{Write named keypairs to an \emph{.Renviron} / \emph{.Rprofile} file}
\usage{
creds_to_renviron(
  ...,
  scope = c("user", "project")[1],
  overwrite = FALSE,
  proj_dir = ".",
  rprofile = FALSE
)
}
\arguments{
\item{...}{named keys to write}

\item{scope}{\code{chr} Edit globally for the current user, or locally for the current project}

\item{overwrite}{\code{(lgl)} should an existing key pair be overwritten. \strong{Default: \code{FALSE}}}

\item{proj_dir}{\code{(chr)} project directory to write credentials to}

\item{rprofile}{\code{(lgl)} whether to write the keypairs to a \link[base]{options} call in a \emph{.Rprofile} file instead.}
}
\value{
success message if a value is written
}
\description{
Writes key pairs to \emph{.Renviron} / \emph{.Rprofile} and adds .Renviron to \emph{.gitignore} if not already there.
}
\seealso{
Other project setup: 
\code{\link{fun_docs_table}()},
\code{\link{ignore_files}()},
\code{\link{install_remote}()},
\code{\link{key_pairs_duplicated}()},
\code{\link{key_pairs_text}()},
\code{\link{use_UU_reimports}()},
\code{\link{use_reimport}()},
\code{\link{write_to_rprofile}()}
}
\concept{project setup}
