% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{css_col2vec}
\alias{css_col2vec}
\title{Convert a CSS representation of a color to an r,g,b numeric}
\usage{
css_col2vec(x)
}
\arguments{
\item{x}{\code{chr} CSS Vector in hexadecimal or rgb/rgba format}
}
\value{
\code{num} with names r,g,b
}
\description{
Convert a CSS representation of a color to an r,g,b numeric
}
\examples{
css_col2vec("#12B4D3")
css_col2vec("rgba(111,96,140,1)")
css_col2vec("green")
}
\seealso{
css_col2vec_
}
