% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{dep_read}
\alias{dep_read}
\title{Read a dependency from file}
\usage{
dep_read(filename, ...)
}
\arguments{
\item{filename}{\code{(chr)} path to the file}

\item{...}{Passed on to read function. See \link[UU]{file_fn} for which function will be used based on the filename}
}
\value{
file contents
}
\description{
Read a dependency from file
}
