% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{dep_write}
\alias{dep_write}
\title{Write a dependency to file}
\usage{
dep_write(x, filename, ...)
}
\arguments{
\item{x}{\code{(object)} to be written}

\item{filename}{\code{(chr)} path to the file}

\item{...}{Passed on to write function. See \link[UU]{file_fn} for which function will be used based on the filename}
}
\value{
\code{(message)} indicating success
}
\description{
Write a dependency to file
}
