% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{dir_fn}
\alias{dir_fn}
\title{Create a directory path pointing function}
\usage{
dir_fn(base_dir)
}
\arguments{
\item{base_dir}{\code{(chr)} the base directory to which the path should point}
}
\value{
\code{(fun)} with pointing function using \link[fs]{path}
}
\description{
Create a directory path pointing function
}
\examples{
dir_fn("data")("random_data", "file", ext = "txt")
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
