% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\docType{data}
\name{dirs}
\alias{dirs}
\title{Path functions for commonly used directories}
\format{
An object of class \code{list} of length 10.
}
\usage{
dirs
}
\arguments{
\item{...}{\code{(chr)} directory paths}

\item{ext}{\code{(chr)} file extension}
}
\description{
Path functions for commonly used directories
}
\examples{
dirs$crmms("crmms_output", "EnsembleOutput_10_mar_2022", ext = "xlsx")
}
\keyword{datasets}
