% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\docType{data}
\name{dirs}
\alias{dirs}
\title{Path functions for commonly used directories}
\format{
An object of class \code{list} of length 13.
}
\usage{
dirs\$data()
}
\arguments{
\item{...}{\code{(chr)} directory paths}

\item{mkpath}{\code{lgl} Whether to return a path regardless of whether the file/dir exists or not}

\item{ext}{\code{(chr)} file extension}

\item{mustWork}{\code{lgl} If \code{TRUE}, an error is given if there are no matching files.}
}
\description{
Path functions for commonly used directories
}
\examples{
dirs$data("mydata", ext = "csv", mkpath = TRUE)
}
\seealso{
Other file IO: 
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
\keyword{datasets}
