% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{ext}
\alias{ext}
\title{Extract the file extensions from a filepath}
\usage{
ext(path, strip = FALSE, new_ext)
}
\arguments{
\item{path}{\code{(character)} path}

\item{strip}{\code{(logical)} Whether to strip the extension from the path to return the bare file name}

\item{new_ext}{\code{chr} New extension for the filename}
}
\value{
\code{(character)} with the extensions
}
\description{
Given a path, extract the file extension
}
\seealso{
Other file IO: 
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
