% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_navigation.R
\docType{data}
\name{file}
\alias{file}
\title{Go to a specified file in the Virga Labs golem project}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
file
}
\arguments{
\item{path}{\code{chr} file to go to. Each method has a preset default based on the function name, but any method can be used to go to any file specified.}
}
\description{
Go to a specified file in the Virga Labs golem project
}
\keyword{datasets}
