% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{file_fn}
\alias{file_fn}
\title{Return the appropriate function for reading the specified path/extension}
\usage{
file_fn(x, write = FALSE)
}
\arguments{
\item{x}{\code{(character)} The extension name or the path to the file}

\item{write}{\code{(logical)} Return the writing function? \strong{Default} \code{FALSE} to return the reading function}
}
\value{
\code{(function)}
}
\description{
Return the appropriate function for reading the specified path/extension
}
\examples{
file_fn("csv")
file_fn("csv", write = TRUE)
}
