% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{find_by_class}
\alias{find_by_class}
\title{Find by class}
\usage{
find_by_class(class, e = rlang::caller_env())
}
\arguments{
\item{class}{The \code{(class)} class to search}

\item{e}{The \code{(environment)} to search}
}
\value{
the first object assigned to the environment that matches the class. If more than one object of the class are found, it triggers a warning.
}
\description{
Find an object by it's class
}
