% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{fml_list}
\alias{fml_list}
\title{Print function formals as a list}
\usage{
fml_list(f, paired = TRUE, to_console = FALSE)
}
\arguments{
\item{f}{\code{fun}}

\item{paired}{\code{lgl} Whether to output as paired arguments to be passed to another function.}

\item{to_console}{\code{lgl} whether to print \code{\link[base]{dput}} output to console. ** Default: FALSE **}
}
\value{
\code{msg}
}
\description{
Print function formals as a list
}
\examples{
fml_list(base::apply)
}
