% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_setup.R
\name{fun_docs_table}
\alias{fun_docs_table}
\title{Create a table of functions and their uses}
\usage{
fun_docs_table(package = pkgload::pkg_name())
}
\arguments{
\item{package}{\code{chr} package name}
}
\value{
\code{shiny.tag}
}
\description{
Create a table of functions and their uses
}
\seealso{
Other project setup: 
\code{\link{creds_to_renviron}()},
\code{\link{ignore_files}()},
\code{\link{install_remote}()},
\code{\link{key_pairs_duplicated}()},
\code{\link{key_pairs_text}()},
\code{\link{use_UU_reimports}()},
\code{\link{use_reimport}()},
\code{\link{write_to_rprofile}()}
}
\concept{project setup}
