% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gbort}
\alias{gbort}
\title{Custom error message
Throw \link[rlang]{abort} with \link[cli]{format_error} capacity on the message}
\usage{
gbort(
  message = NULL,
  class = NULL,
  ...,
  trace = rlang::trace_back(),
  parent = NULL,
  e = rlang::caller_env()
)
}
\arguments{
\item{message}{The message to display. Character vectors are
formatted with \code{\link[rlang:format_error_bullets]{format_error_bullets()}}. The first element
defines a message header and the rest of the vector defines
bullets. Bullets named \code{i} and \code{x} define info and error bullets
respectively, with special Unicode and colour formatting applied
if possible.

If a message is not supplied, it is expected that the message is
generated lazily through \link[rlang:cnd_message]{conditionMessage()}. In
that case, \code{class} must be supplied. Only \code{inform()} allows empty
messages as it is occasionally useful to build user output
incrementally.}

\item{class}{Subclass of the condition. This allows your users
to selectively handle the conditions signalled by your functions.}

\item{...}{Additional data to be stored in the condition object.}

\item{trace}{A \code{trace} object created by \code{\link[rlang:trace_back]{trace_back()}}.}

\item{parent}{A parent condition object created by \code{\link[rlang:abort]{abort()}}.}

\item{e}{\code{(environment)} calling environment. Passed to \code{glue} for making the message}
}
\description{
Custom error message
Throw \link[rlang]{abort} with \link[cli]{format_error} capacity on the message
}
