% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gbort}
\alias{gbort}
\title{Custom error message}
\usage{
gbort(
  message = NULL,
  class = NULL,
  ...,
  trace = rlang::trace_back(bottom = 2),
  parent = NULL,
  e = rlang::caller_env()
)
}
\arguments{
\item{message}{\code{(chr)} The message to display, formatted as a bulleted list. The first element is displayed as an alert bullet prefixed with ! by default. Elements named "*", "i", and "x" are formatted as regular, info, and cross bullets respectively. See \link[rlang]{topic-condition-formatting} for more about bulleted messaging.}

\item{class}{\code{(class)} Subclass of the condition}

\item{trace}{\code{(trace)} A \code{trace} object created by \link[rlang]{trace_back}}

\item{parent}{\code{(cond)} Supply \code{parent} when you rethrow an error from a condition handler}

\item{e}{\code{(environment)} calling environment. Passed to \code{glue} for making the message}
}
\description{
Throw \link[rlang]{abort} with \link[cli]{format_error}
}
\seealso{
Other condition signaling: 
\code{\link{gmsg}()},
\code{\link{gwarn}()}
}
\concept{condition signaling}
