% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{glue_js}
\alias{glue_js}
\title{Create a JS string with glue insertions
glue \verb{.open = !@} & \verb{.close = @#}}
\usage{
glue_js(
  js,
  as_chr = FALSE,
  e = rlang::caller_env(),
  .open = "*{",
  .close = "}*"
)
}
\arguments{
\item{js}{\code{chr} JS code to \link[glue]{glue}}

\item{as_chr}{\code{lgl} Whether to use \code{\link[UU]{as_js}} on the output \code{FALSE} or \link[base]{as.character} \code{TRUE}. \strong{Default FALSE}}

\item{e}{\code{env} calling environment}
}
\value{
\code{chr}
}
\description{
Create a JS string with glue insertions
glue \verb{.open = !@} & \verb{.close = @#}
}
\examples{
glue_js("$(document).ready(() => {let x = *{tolower(FALSE)}*)")
}
\seealso{
Other shiny: 
\code{\link{as_js}()},
\code{\link{nm_to_id}()},
\code{\link{read_js}()},
\code{\link{shiny_error_recover}()},
\code{\link{strip_html}()}
}
\concept{shiny}
