% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{glue_js}
\alias{glue_js}
\title{Create a JS string with glue insertions
glue \verb{.open = !@} & \verb{.close = @#}}
\usage{
glue_js(js, e = rlang::caller_env(), .open = "*{", .close = "}*")
}
\arguments{
\item{js}{\code{chr} JS code to \link[glue]{glue}}

\item{e}{\code{env} calling environment}
}
\value{
\code{chr}
}
\description{
Create a JS string with glue insertions
glue \verb{.open = !@} & \verb{.close = @#}
}
\examples{
gluejs("$(document).ready(() => {let x = *{tolower(F)}*)")
}
