% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gmessage}
\alias{gmessage}
\title{Custom warning message
Throw \link[rlang]{signal} with \link[cli]{format_message} capacity on the message}
\usage{
gmessage(message = NULL, class = "message", ...)
}
\arguments{
\item{message}{The message to display. Character vectors are
formatted with \code{\link[rlang:format_error_bullets]{format_error_bullets()}}. The first element
defines a message header and the rest of the vector defines
bullets. Bullets named \code{i} and \code{x} define info and error bullets
respectively, with special Unicode and colour formatting applied
if possible.

If a message is not supplied, it is expected that the message is
generated lazily through \link[rlang:cnd_message]{conditionMessage()}. In
that case, \code{class} must be supplied. Only \code{inform()} allows empty
messages as it is occasionally useful to build user output
incrementally.}

\item{class}{Subclass of the condition. This allows your users
to selectively handle the conditions signalled by your functions.}

\item{...}{Additional data to be stored in the condition object.}
}
\description{
Custom warning message
Throw \link[rlang]{signal} with \link[cli]{format_message} capacity on the message
}
