% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gmsg}
\alias{gmsg}
\title{Custom message
Message using \link[cli]{format_message} & \link[cli]{cat_line}}
\usage{
gmsg(msg, e = rlang::caller_env())
}
\description{
Custom message
Message using \link[cli]{format_message} & \link[cli]{cat_line}
}
\seealso{
Other condition signaling: 
\code{\link{gbort}()},
\code{\link{ginfo}()},
\code{\link{gwarn}()}
}
\concept{condition signaling}
