% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infixes.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Are lhs values absent from set on rhs?}
\usage{
... \%nin\% NA
}
\description{
Are lhs values absent from set on rhs?
}
\examples{
1 \%nin\% 1:10
}
\seealso{
is_legit

Other infixes: 
\code{\link{\%allin\%}()},
\code{\link{\%|0|\%}()},
\code{\link{\%|\%}()},
\code{\link{\%|legit|\%}()},
\code{\link{\%|try|\%}()},
\code{\link{\%|zchar|\%}()}
}
\author{
Think.fr
}
\concept{infixes}
