% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infixes.R
\name{\%|legit|\%}
\alias{\%|legit|\%}
\title{If legit lhs, else rhs}
\usage{
lhs \%|legit|\% rhs
}
\arguments{
\item{lhs}{\code{(expression)} to try}

\item{rhs}{\code{(expression)} to replace if expression fails}
}
\value{
If legit lhs else rhs
}
\description{
If legit lhs, else rhs
}
\examples{
(100 / NA) \%|legit|\% 4
list(a = 5)$a \%|legit|\% 4
list(a = 5)$b \%|legit|\% 4
}
\seealso{
is_legit
}
