% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{\%|try|\%}
\alias{\%|try|\%}
\title{Try an expression}
\usage{
lhs \%|try|\% rhs
}
\arguments{
\item{lhs}{\code{(expression)} to try}

\item{rhs}{\code{()}}
}
\value{
rhs
}
\description{
Calls the expression (LHS) & if it fails return RHS
}
