% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{gwarn}
\alias{gwarn}
\title{Custom warning message}
\usage{
gwarn(
  message = NULL,
  body = NULL,
  footer = NULL,
  .frequency = c("always", "regularly", "once"),
  e = rlang::caller_env()
)
}
\arguments{
\item{message}{\code{(chr)} The message to display, formatted as a bulleted list. The first element is displayed as an alert bullet prefixed with ! by default. Elements named "*", "i", and "x" are formatted as regular, info, and cross bullets respectively. See \link[rlang]{topic-condition-formatting} for more about bulleted messaging.}

\item{body, footer}{Additional bullets.}

\item{.frequency}{How frequently should the warning or message be
displayed? By default (\code{"always"}) it is displayed at each
time. If \code{"regularly"}, it is displayed once every 8 hours. If
\code{"once"}, it is displayed once per session.}

\item{e}{\code{(environment)} calling environment. Passed to \code{glue} for making the message}
}
\description{
Throw \link[cli]{cli_alert_warning} with \link[cli]{format_warning}
}
\seealso{
Other condition signaling: 
\code{\link{gbort}()},
\code{\link{gmsg}()}
}
\concept{condition signaling}
