% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_helpers.R
\name{if_debug}
\alias{if_debug}
\title{Run expressions only when option \code{use_debug = TRUE}}
\usage{
if_debug(...)
}
\arguments{
\item{...}{\code{expr} to evaluate}
}
\description{
Run expressions only when option \code{use_debug = TRUE}
}
