% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{ignore_files}
\alias{ignore_files}
\title{Add lines to \emph{.gitignore}}
\usage{
ignore_files(lines, directory = ".")
}
\arguments{
\item{lines}{\code{chr}}

\item{directory}{\code{(chr)} directory path of \emph{.gitignore} to be modified}
}
\value{
\code{informative messages}
}
\description{
Add lines to \emph{.gitignore}
}
