% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{interpolate}
\alias{interpolate}
\title{Simple interpolate between two numbers}
\usage{
interpolate(start, end, n)
}
\arguments{
\item{start}{\code{num} first number}

\item{end}{\code{num} last number}

\item{n}{\code{int} total numbers in output vector}
}
\value{
\code{num} vector with length equivalent to n
}
\description{
Simple interpolate between two numbers
}
\examples{
interpolate(0, 6, 6)
}
