% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{is_error}
\alias{is_error}
\title{Is object an error class?}
\usage{
is_error(x)
}
\arguments{
\item{x}{\code{(object)}}
}
\value{
\code{(logical)}
}
\description{
Is object of class \code{try-error}
}
