% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{is_error}
\alias{is_error}
\title{Is object an error}
\usage{
is_error(x)
}
\arguments{
\item{x}{\code{obj}}
}
\value{
\code{lgl}
}
\description{
Is object an error
}
\examples{
is_error(try(stop()))
}
