% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{is_filepath}
\alias{is_filepath}
\title{Is path a file path}
\usage{
is_filepath(path)
}
\arguments{
\item{path}{\code{(character)} path}
}
\value{
\code{(logical)}
}
\description{
Given a path, is it a filepath?
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{load_obj}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
