% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditionals.R
\name{is_legit}
\alias{is_legit}
\title{Is object legit?}
\usage{
is_legit(x)
}
\arguments{
\item{x}{\code{(object)}}
}
\value{
\code{(logical)}
}
\description{
Is object non-null, non-empty, non-NA, and not a try-error?
}
\seealso{
Other conditionals: 
\code{\link{is_error}()},
\code{\link{is_project}()},
\code{\link{larger}()},
\code{\link{most}()},
\code{\link{same}()},
\code{\link{zchar}()}
}
\concept{conditionals}
