% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{is_legit}
\alias{is_legit}
\title{Is object legit?}
\usage{
is_legit(x)
}
\arguments{
\item{x}{\code{(object)}}
}
\value{
\code{(logical)}
}
\description{
Is object non-null, non-empty, non-NA, and not a try-error?
}
