% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_helpers.R
\name{is_package}
\alias{is_package}
\title{Is working directory a package?}
\usage{
is_package()
}
\value{
\code{lgl}
}
\description{
Is working directory a package?
}
\examples{
is_package()
}
\seealso{
Other package dev: 
\code{\link{get_package_fns}()},
\code{\link{is_package_dev}()}
}
\concept{package dev}
