% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_helpers.R
\name{is_package_dev}
\alias{is_package_dev}
\title{Is package in development or installed}
\usage{
is_package_dev(pkg_nm = pkgload::pkg_name() \%|try|\% NULL)
}
\arguments{
\item{pkg_nm}{\code{chr} name of package}
}
\value{
\code{lgl}
}
\description{
Is package in development or installed
}
