% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{key_pairs_text}
\alias{key_pairs_text}
\title{Make key-pairs from a named character vector}
\usage{
key_pairs_text(x)
}
\arguments{
\item{x}{\code{chr} named}
}
\value{
\code{chr}
}
\description{
Make key-pairs from a named character vector
}
\examples{
key_pairs_text(Sys.getenv())
}
