% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{last_updated}
\alias{last_updated}
\title{Gather last updated times for on-disk files}
\usage{
last_updated(x, path = FALSE)
}
\arguments{
\item{x}{\code{(chr)} file path to check last updated time}

\item{path}{\code{(lgl)} whether x is a path and all files should be checked}
}
\value{
\code{(POSIXct)} Last modified time
}
\description{
Check the last modified time files or paths
}
