% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{len_unique}
\alias{len_unique}
\title{The length of unique values in a vector}
\usage{
len_unique(x)
}
\arguments{
\item{x}{\code{vctr}}
}
\value{
\code{dbl}
}
\description{
The length of unique values in a vector
}
\examples{
len_unique(c(1,2,1))
}
