% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{magnitude_order}
\alias{magnitude_order}
\title{Compute the order of magnitude}
\usage{
magnitude_order(x)
}
\arguments{
\item{x}{\code{num}}
}
\value{
\code{num}
}
\description{
Uses the \link[base]{floor} to round
}
\examples{
magnitude_order(10^(1:10))
}
