% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{match_df}
\alias{match_df}
\title{Extract matching rows of a data frame.}
\usage{
match_df(x, y, out = NULL, on = NULL, verbose = FALSE)
}
\arguments{
\item{x}{data frame to subset.}

\item{y}{data frame defining matching rows.}

\item{out}{\code{obj} Of class matching the desired output. \strong{Default} \code{NULL} returns a \code{data.frame} with the row(s) of \code{x} that have matches in \code{y}. \code{numeric()} will return the matching indices of \code{x} with matches in \code{y} & \code{logical()} will return a matching logical vector with length equivalent to \code{x} of the rows matching in \code{y}}

\item{on}{\code{chr} Either a vector of the names on which to match, if they are named similarly in x & y, or a specification in the form \code{c(y_feature = x_feature)}}
}
\value{
\code{tbl/dbl/lgl} Depending on the class of \code{out}
}
\description{
Match works in the same way as join, but instead of return the combined
dataset, it only returns the matching rows from the first dataset. This is
particularly useful when you've summarised the data in some way
and want to subset the original data by a characteristic of the subset.
}
\seealso{
plyr::match_df
}
