% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{missing_args}
\alias{missing_args}
\title{Get the missing arguments from the function as character}
\usage{
missing_args(
  calling_function = rlang::caller_fn(1),
  corresponding_call = sys.call(1),
  include_null = TRUE,
  exclude_defaults = TRUE
)
}
\arguments{
\item{calling_function}{\code{(function)} see \link[rlang]{caller_fn} or \link[base]{sys.function}}

\item{corresponding_call}{\code{(call)} The call where the \code{calling_function} is called. See \link[rlang]{trace_back} or \link[base]{sys.call}}

\item{include_null}{\code{(logical)} Include args set to \code{NULL}?}

\item{exclude_defaults}{\code{(logical)} Exclude arguments wth defaults?}
}
\value{
\code{(character)}
}
\description{
Get the missing arguments from the function as character
}
\examples{
a <- function(a, b = NULL, c = "d") {
  missing_args()
}
a()
}
