% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{mkpath}
\alias{mkpath}
\title{Construct a path}
\usage{
mkpath(path, mkfile = FALSE, mkdir = TRUE)
}
\arguments{
\item{path}{\code{(character)} path}

\item{mkfile}{\code{(logical)} whether to make the file if it doesn't exist. IF \code{TRUE} and the path has an extension, both the directory and the file will be created}

\item{mkdir}{\code{(logical)} whether to make the directory if it doesn't exist. If \code{TRUE}, and the \code{path} does not have an extension, path will be created as a directory path.}
}
\value{
\code{(informative messages)}
}
\description{
Given a path, construct it if it does not exist.
}
