% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{mkpath}
\alias{mkpath}
\title{Construct a path}
\usage{
mkpath(path, mkfile = FALSE)
}
\arguments{
\item{path}{\code{(character)} path}

\item{mkfile}{\code{(logical)} whether to make the file if it doesn't exist.}
}
\value{
\code{(informative messages)}
}
\description{
Given a path, construct it if it does not exist.
}
