% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{mkpath}
\alias{mkpath}
\title{Construct a path}
\usage{
mkpath(path, mkfile = FALSE, mkdir = TRUE)
}
\arguments{
\item{path}{\code{(character)} path}

\item{mkfile}{\code{(logical)} whether to make the file if it doesn't exist. IF \code{TRUE} and the path has an extension, both the directory and the file will be created}

\item{mkdir}{\code{(logical)} whether to make the directory if it doesn't exist. If \code{TRUE}, and the \code{path} does not have an extension, path will be created as a directory path.}
}
\value{
\code{(informative messages)}
}
\description{
Given a path, construct it if it does not exist.
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
