% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vectors.R
\name{names_values_switch}
\alias{names_values_switch}
\title{Switch the names and the values of a vector}
\usage{
names_values_switch(x)
}
\arguments{
\item{x}{\code{named object}}
}
\value{
\code{obj}
}
\description{
Switch the names and the values of a vector
}
\examples{
names_values_switch(c(a = 1, b = 2))
}
\seealso{
Other vectors: 
\code{\link{len_unique}()},
\code{\link{rle_df}()},
\code{\link{rle_groups}()},
\code{\link{rle_seq}()},
\code{\link{sort_by_names}()},
\code{\link{true_names}()},
\code{\link{unify_vec_preserve_order}()},
\code{\link{unique_with_names}()},
\code{\link{vlookup}()},
\code{\link{vlookup_from_ref}()},
\code{\link{zchar_remove}()}
}
\concept{vectors}
