% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{needs_update}
\alias{needs_update}
\title{Check if files need to be updated}
\usage{
needs_update(
  x,
  path = FALSE,
  threshold = lubridate::floor_date(Sys.time(), "day")
)
}
\arguments{
\item{x}{\code{(chr)} file path to check last updated time}

\item{path}{\code{(lgl)} whether x is a path and all files should be checked}

\item{threshold}{The threshold time. If files have last modified times less than this time, they will be marked as needing an update.}
}
\value{
\code{(tbl)} with columns:
\itemize{
\item{\code{full_path}}{ The full path to the file(s)}
\item{\code{basename}}{ The file(s) basename}
\item{\code{last_updated}}{ The last updated time}
\item{\code{threshold}}{ The threshold time for comparison}
\item{\code{needs_update}}{ logical as to whether the file should be updated}
}
}
\description{
Check if files need to be updated
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
