% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rounding.R
\docType{data}
\name{num_chr_suffi}
\alias{num_chr_suffi}
\title{Abbreviations of numeric magnitude}
\format{
An object of class \code{character} of length 4.
}
\usage{
num_chr_suffi
}
\description{
Abbreviations of numeric magnitude
}
\seealso{
Other rounding: 
\code{\link{magnitude_order}()},
\code{\link{magnitude_triplet}()},
\code{\link{num2str}()},
\code{\link{num2str_vec}()},
\code{\link{round_to}()},
\code{\link{size}()},
\code{\link{unit_find}()},
\code{\link{unit_modify}()},
\code{\link{unit_modify_vec}()},
\code{\link{unit_string}()}
}
\concept{rounding}
\keyword{datasets}
