% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{object_fn}
\alias{object_fn}
\title{Return the appropriate function for writing the supplied object to disk}
\usage{
object_fn(x, filepath)
}
\arguments{
\item{x}{\code{(object)}}
}
\value{
\code{(function)} See details for which function
}
\description{
Return the appropriate function for writing the supplied object to disk
}
\details{
\itemize{
\item{\code{data.frame/matrix}}{ \link[arrow]{write_feather}}
\item{\code{ggplot}}{ \link[ggplot2]{ggsave}}
\item{\code{anything else}}{ \link[base]{saveRDS}}
}
}
\examples{
object_fn(1:15)
object_fn(data.frame(a = 2, b = 3))
}
