% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{object_fn}
\alias{object_fn}
\title{Return the appropriate function for writing the supplied object to disk}
\usage{
object_fn(x, filepath)
}
\arguments{
\item{x}{\code{(object)}}
}
\value{
\code{(function)}
}
\description{
Return the appropriate function for writing the supplied object to disk
}
\examples{
object_fn(1:15)
object_fn(data.frame(a = 2, b = 3))
}
