% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{object_write}
\alias{object_write}
\title{Provide the appropriate file read/write function}
\usage{
object_write(x, filename, path, ..., verbose = TRUE)
}
\arguments{
\item{x}{\code{(object)} to write to disk}

\item{filename}{\code{(chr)} without path to write to disk}

\item{path}{\code{(chr)} where file will be written}

\item{...}{arguments passed on to methods. See \code{?object_fn}}

\item{verbose}{\code{(logical)} Whether to print saved messages. \strong{Default} \code{TRUE}}
}
\value{
Success message if file is written
}
\description{
Write an object to disk
}
