% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{object_write}
\alias{object_write}
\title{Provide the appropriate file read/write function}
\usage{
object_write(x, filename, path = ".", ..., verbose = TRUE)
}
\arguments{
\item{x}{\code{(object)} to write to disk}

\item{filename}{\code{(chr)} without path to write to disk}

\item{path}{\code{(chr)} where file will be written}

\item{...}{arguments passed on to methods. See \code{?object_fn}}

\item{verbose}{\code{(logical)} Whether to print saved messages. \strong{Default} \code{TRUE}}
}
\value{
Success message if file is written
}
\description{
Write an object to disk
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()},
\code{\link{write_lines}()}
}
\concept{file IO}
