% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts_helpers.R
\docType{data}
\name{opts}
\alias{opts}
\title{Check option value. \strong{Use interactively only!}}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
opts
}
\arguments{
\item{default}{if the specified option is not set in the options list, this value is returned. This facilitates retrieving an option and checking whether it is set and setting it separately if not.}
}
\value{
\code{lgl}
}
\description{
Check option value. \strong{Use interactively only!}
}
\seealso{
write_opts
Code is dynamically generated when package is loaded. These will not auto generate properly in a deployed shiny app. See \link{write_opts} for a non-interactive alternative.

Other options: 
\code{\link{if_debug}()},
\code{\link{toggle}},
\code{\link{write_opts}()}
}
\concept{options}
\keyword{datasets}
