% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{package_size}
\alias{package_size}
\title{Return the size of a package, or all packages in a folder}
\usage{
package_size(packages, path = .libPaths()[1])
}
\arguments{
\item{packages}{\code{chr} of package names}

\item{path}{\code{chr} with package folders in it such as \code{\link[base]{.libPaths}}}
}
\value{
\code{chr/tbl} depending on whether packages or path is provided
}
\description{
Return the size of a package, or all packages in a folder
}
\examples{
package_size()
}
