% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environments_namespaces.R
\name{pkg_ns}
\alias{pkg_ns}
\title{Return the current package namespace}
\usage{
pkg_ns(pkg = basename(rprojroot::find_package_root_file()))
}
\arguments{
\item{pkg}{\code{chr} package name. Defaults to current}
}
\value{
\code{env}
}
\description{
Return the current package namespace
}
\seealso{
Other namespaces: 
\code{\link{assign_global}()},
\code{\link{assign_in_ns}()},
\code{\link{get_from_ns}()},
\code{\link{get_global}()},
\code{\link{unload_namespaces}()}
}
\concept{namespaces}
