% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{regex_op}
\alias{regex_op}
\title{Create a compound regex grouped statement}
\usage{
regex_op(x, type = "|")
}
\arguments{
\item{x}{\code{(character)} regex strings}

\item{type}{\code{(character)} \code{|}, \code{&} supported}
}
\value{
\code{(character)} grouped regex statement
}
\description{
Create a compound regex grouped statement
}
