% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{regex_or}
\alias{regex_or}
\title{Create a compound regex grouped OR statement}
\usage{
regex_or(x)
}
\arguments{
\item{x}{\code{(character)} regex strings}
}
\value{
\code{(character)} grouped regex OR statement
}
\description{
Create a compound regex grouped OR statement
}
