% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_color_helpers.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert r,g,b,a values as string or numeric to hex}
\usage{
rgb2hex(red, green, blue, with_alpha = FALSE)
}
\arguments{
\item{red}{\code{chr/num} Either a CSS \code{rgb()} or \code{rgba()} declaration as a string, or the red value as numeric.}

\item{with_alpha}{\code{lgl} Whether to include the alpha value (an 8 digit hex) in the output or not}
}
\value{
\code{chr} The hex value
}
\description{
Convert r,g,b,a values as string or numeric to hex
}
\examples{
rgb2hex("rgba(18,180,211,1)", with_alpha = TRUE)
rgb2hex("rgba(18,180,211,1)", with_alpha = FALSE)
}
