% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{rle_df}
\alias{rle_df}
\title{rle_df

Given an \code{\link[base]{rle}} this function will return a data.frame of starts, ends, and indexes thereof of the run lengths.
Credit: \url{https://stackoverflow.com/questions/43875716/find-start-and-end-positions-indices-of-runs-consecutive-values}}
\usage{
rle_df(x)
}
\arguments{
\item{x}{\code{(vector)} An object for which to run an \code{rle}}
}
\value{
\item{(data.frame)}{ with length, values, start and end indices.}
}
\description{
rle_df

Given an \code{\link[base]{rle}} this function will return a data.frame of starts, ends, and indexes thereof of the run lengths.
Credit: \url{https://stackoverflow.com/questions/43875716/find-start-and-end-positions-indices-of-runs-consecutive-values}
}
\examples{
rle_df(sample(c(TRUE,FALSE), replace = TRUE, 100))
}
