% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{rle_groups}
\alias{rle_groups}
\title{Create an RLE Grouping from a logical vector}
\usage{
rle_groups(x)
}
\arguments{
\item{x}{\code{lgl} vector}
}
\value{
\code{list}
}
\description{
Create an RLE Grouping from a logical vector
}
\examples{
rle_groups(sample(c(TRUE, FALSE), 20, replace = TRUE))
}
