% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{rle_seq}
\alias{rle_seq}
\title{Create a sequence from the start to the end for a given value from an \code{rle_df} for indexing}
\usage{
rle_seq(rle_df, value)
}
\arguments{
\item{rle_df}{\code{(tbl)} See \code{rle_df}}

\item{value}{\code{(any)} Value to filter for in the \code{values} column. Require the values in the value column to be unique.}
}
\value{
\code{(dbl)}
}
\description{
Create a sequence from the start to the end for a given value from an \code{rle_df} for indexing
}
\examples{
rle_seq(rle_df(rep(letters[1:3], each = 3)), "c")
}
