% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{round_to}
\alias{round_to}
\title{Find the global minima/maxima of input vectors}
\usage{
round_to(..., accuracy = NULL, fn = min, f = NULL)
}
\arguments{
\item{...}{\code{num} vectors}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{fn}{\code{fun} \link[base]{min}/\link[base]{max}}

\item{f}{rounding function: \code{\link{floor}}, \code{\link{ceiling}} or
\code{\link{round}}}
}
\value{
\code{num}
}
\description{
If accuracy is omitted, number will be rounded to the nearest order of magnitude IE 145, if \code{fn = min}, will round to 100
}
\examples{
round_to(runif(10, 5:10), runif(2, 2:5))
round_to(145)
}
\seealso{
Other rounding: 
\code{\link{magnitude_order}()},
\code{\link{magnitude_triplet}()},
\code{\link{num2str_vec}()},
\code{\link{num2str}()},
\code{\link{num_chr_suffi}},
\code{\link{size}()},
\code{\link{unit_find}()},
\code{\link{unit_modify_vec}()},
\code{\link{unit_modify}()},
\code{\link{unit_string}()}
}
\concept{rounding}
