% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditionals.R
\name{same}
\alias{same}
\title{Are the values in each object the same?}
\usage{
same(target, current, sort_by_names = TRUE, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[base:all.equal]{base::all.equal}}
  \describe{
    \item{\code{target}}{\R object.}
    \item{\code{current}}{other \R object, to be compared with \code{target}.}
  }}

\item{x}{\code{obj}}

\item{y}{\code{obj}}
}
\value{
\code{lgl}
}
\description{
The primary difference from \code{\link[base]{identical}} & \code{\link[base]{all.equal}} is that objects are sorted by name so order doesn't matter. Set \code{sort_by_names = FALSE} to sort by values.
}
\examples{
same(list(x = 1, y = 2), list(y = 2, x = 1))
}
\seealso{
Other conditionals: 
\code{\link{is_error}()},
\code{\link{is_legit}()},
\code{\link{is_project}()},
\code{\link{larger}()},
\code{\link{most}()},
\code{\link{zchar}()}
}
\concept{conditionals}
