% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{same}
\alias{same}
\title{Are the values in each object the same?}
\usage{
same(x, y)
}
\arguments{
\item{x}{\code{obj}}

\item{y}{\code{obj}}
}
\value{
\code{lgl}
}
\description{
The primary difference from \code{\link[base]{identical}} & \code{\link[base]{all.equal}} is that objects are sorted by name so order doesn't matter.
}
\examples{
same(list(x = 1, y = 2), list(y = 2, x = 1))
}
