% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{shiny_error_recover}
\alias{shiny_error_recover}
\title{Toggle \link[utils]{recover} on error when obtuse shiny errors are encountered}
\usage{
shiny_error_recover()
}
\description{
Toggle \link[utils]{recover} on error when obtuse shiny errors are encountered
}
\seealso{
Other shiny: 
\code{\link{as_js}()},
\code{\link{glue_js}()},
\code{\link{nm_to_id}()},
\code{\link{path_strip_shiny}()},
\code{\link{path_strip_to}()},
\code{\link{read_js}()},
\code{\link{strip_html}()}
}
\concept{shiny}
