% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{size}
\alias{size}
\alias{size_}
\title{Digital storage size conversion
See \link[utils]{object.size}}
\usage{
size(
  x,
  in_unit = "b",
  out_unit = c("b", "Kb", "Mb", "Gb", "Tb", "Pb", "Eb", "Zb", "Yb")[3]
)

size_(
  x,
  in_unit = "b",
  out_unit = c("b", "Kb", "Mb", "Gb", "Tb", "Pb", "Eb", "Zb", "Yb")[3]
)
}
\arguments{
\item{x}{\code{(numeric)}}

\item{in_unit}{\code{(character)} units of x}

\item{out_unit}{\code{(character)} units of output number \strong{Default: Mb}. See \link[utils]{object.size} for the options.}

\item{standard}{\code{(character)}}
}
\value{
\code{(numeric)} value in \code{out_unit}s
}
\description{
Digital storage size conversion
See \link[utils]{object.size}

Vectorized version of size
}
\examples{
size(50, "Mb")
size(50, "Gb")
size(50, "Gb", "Mb")
}
\seealso{
size_

Other rounding: 
\code{\link{magnitude_order}()},
\code{\link{magnitude_triplet}()},
\code{\link{num2str_vec}()},
\code{\link{num2str}()},
\code{\link{num_chr_suffi}},
\code{\link{round_to}()},
\code{\link{unit_find}()},
\code{\link{unit_modify_vec}()},
\code{\link{unit_modify}()},
\code{\link{unit_string}()}
}
\concept{rounding}
