% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vectors.R
\name{sort_by_names}
\alias{sort_by_names}
\title{Sort a vector or list by it's name (or self if no names)}
\usage{
sort_by_names(x, by_names = TRUE, sort_by_values = TRUE)
}
\arguments{
\item{x}{\code{obj} to sort}

\item{by_names}{\code{lgl} whether to sort by names}

\item{sort_by_values}{\code{lgl} whether to sort by values in the absence of names. If FALSE and \code{x} has no names, an error will be thrown.}
}
\value{
\code{obj} sorted
}
\description{
Sort a vector or list by it's name (or self if no names)
}
\examples{
sort_by_names(c(b = "b", c = "a"))
sort_by_names(c(b = "b", c = "a"), by_names = FALSE)
}
\seealso{
Other vectors: 
\code{\link{len_unique}()},
\code{\link{names_values_switch}()},
\code{\link{rle_df}()},
\code{\link{rle_groups}()},
\code{\link{rle_seq}()},
\code{\link{true_names}()},
\code{\link{unify_vec_preserve_order}()},
\code{\link{unique_with_names}()},
\code{\link{vlookup}()},
\code{\link{vlookup_from_ref}()},
\code{\link{zchar_remove}()}
}
\concept{vectors}
