% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{str_break_every}
\alias{str_break_every}
\title{Break word every x characters}
\usage{
str_break_every(x, every)
}
\arguments{
\item{x}{\code{chr} vector}

\item{every}{\code{num} of chars between each line break}
}
\value{
\code{chr}
}
\description{
Break word every x characters
}
\examples{
str_break_every("I am a very long string that will broken into pieces", 10)
}
