% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{strip_html}
\alias{strip_html}
\title{Remove all HTML tags from a character vector}
\usage{
strip_html(x)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{chr} without HTML tags
}
\description{
Remove all HTML tags from a character vector
}
\examples{
strip_html("The <strong>fox</strong> ran")
}
\seealso{
Other shiny: 
\code{\link{as_js}()},
\code{\link{glue_js}()},
\code{\link{nm_to_id}()},
\code{\link{path_strip_shiny}()},
\code{\link{path_strip_to}()},
\code{\link{read_js}()},
\code{\link{shiny_error_recover}()}
}
\concept{shiny}
