% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\docType{data}
\name{time_aggregates}
\alias{time_aggregates}
\title{Timespans as character}
\format{
An object of class \code{list} of length 7.
}
\usage{
time_aggregates
}
\description{
Timespans as character
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{excel_date}()},
\code{\link{month_factor}()},
\code{\link{season_factor}()},
\code{\link{time_difftimes}},
\code{\link{time_elapsed}()},
\code{\link{time_factor}()},
\code{\link{timespan}()},
\code{\link{week_factor}()}
}
\concept{time}
\keyword{datasets}
