% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\docType{data}
\name{time_difftimes}
\alias{time_difftimes}
\title{Timespans as durations}
\format{
An object of class \code{list} of length 7.
}
\usage{
time_difftimes
}
\description{
Timespans as durations
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{excel_date}()},
\code{\link{month_factor}()},
\code{\link{season_factor}()},
\code{\link{time_aggregates}},
\code{\link{time_elapsed}()},
\code{\link{time_factor}()},
\code{\link{timespan}()},
\code{\link{week_factor}()}
}
\concept{time}
\keyword{datasets}
