% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{time_elapsed}
\alias{time_elapsed}
\title{Return a logical on an interval}
\usage{
time_elapsed(file = ".interval_timer.rds", interval = lubridate::weeks(1))
}
\arguments{
\item{file}{\code{chr} filename in which to store the interval time}

\item{interval}{\code{period/duration} Default 1 week}
}
\value{
\code{lgl}
}
\description{
Return a logical on an interval
}
\examples{
time_elapsed()
}
