% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\name{time_factor}
\alias{time_factor}
\title{Turn timespans into an ordered factor}
\usage{
time_factor(x)
}
\arguments{
\item{x}{\code{chr} of timespans}
}
\value{
\code{factor}
}
\description{
Turn timespans into an ordered factor
}
\examples{
time_factor(c("year", "week"))
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{excel_date}()},
\code{\link{season_factor}()},
\code{\link{time_aggregates}},
\code{\link{time_difftimes}},
\code{\link{time_elapsed}()},
\code{\link{timespan}()},
\code{\link{week_factor}()}
}
\concept{time}
