% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\name{timespan}
\alias{timespan}
\title{Create a timespan duration}
\usage{
timespan(x)
}
\arguments{
\item{x}{\code{chr/POSIXt/Date}
\itemize{
\item{\code{chr}}{ One of year, season, quarter, month, week, day, hour}
\item{\code{POSIXt/Date/Datetime}}{ A Date or Datetime vector}
}}
}
\value{
\code{Duration} See \code{\link[lubridate]{duration}}
}
\description{
Create a timespan duration
}
\examples{
timespan(seq(Sys.time() - lubridate::dyears(), Sys.time(), by = 1))
timespan("season")
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{excel_date}()},
\code{\link{month_factor}()},
\code{\link{season_factor}()},
\code{\link{time_aggregates}},
\code{\link{time_difftimes}},
\code{\link{time_elapsed}()},
\code{\link{time_factor}()},
\code{\link{week_factor}()}
}
\concept{time}
