% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts_helpers.R
\docType{data}
\name{toggle}
\alias{toggle}
\title{Toggle or change an option listed in a local \emph{.Rprofile} for the session}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
toggle$your_logical_option_name()

toggle$your_non_logical_option_name(set = "value")
}
\arguments{
\item{opt}{\code{chr} option name to toggle. Dynamically populated}

\item{set}{\code{obj} value to set the option to (takes precedence over toggle)}
}
\value{
\code{msg} to inform user of what the option is set to
}
\description{
Any options in the project local \emph{.Rprofile} will populate this object as named methods. These named methods, when called, will toggle the options on or off. Alternatively, if an option is not logical, it can be changed using the \code{set} argument. This loads at the start of an R session if \link[UU]{startup} has been called in the user-level \emph{.Rprofile}. These methods are meant for interactive use only.
}
\seealso{
Other options: 
\code{\link{if_debug}()},
\code{\link{opts}},
\code{\link{write_opts}()}
}
\concept{options}
\keyword{datasets}
