% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conditionals.R
\name{true_names}
\alias{true_names}
\title{Return the names of all TRUE items in a logical vector}
\usage{
true_names(x)
}
\arguments{
\item{x}{\code{lgl} with names}
}
\value{
\code{chr} of the names of true values
}
\description{
Return the names of all TRUE items in a logical vector
}
\examples{
true_names(c(a = TRUE, b = FALSE))
true_names(c(a = FALSE, b = FALSE))
}
