% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unify_vec_preserve_order}
\alias{unify_vec_preserve_order}
\title{Unify two vectors preserving the order of \code{x}}
\usage{
unify_vec_preserve_order(x, y)
}
\arguments{
\item{x}{\code{vec} to preserve the order of}

\item{y}{\code{vec} to vector of values to include in the output (unordered)}
}
\description{
Unify two vectors preserving the order of \code{x}
}
\examples{
unify_vec_preserve_order(letters[c(5, 3)], letters[c(4:10,3)])
unify_vec_preserve_order(letters[1:5], letters[c(4:10)])
unify_vec_preserve_order(NULL, letters[c(4:10)])
unify_vec_preserve_order(letters, NULL)
}
