% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vectors.R
\name{unique_with_names}
\alias{unique_with_names}
\title{Unique a vector, preserving the names of the first original entries}
\usage{
unique_with_names(x)
}
\arguments{
\item{x}{\code{vec} with names}
}
\value{
\code{vec} of the same type
}
\description{
Unique a vector, preserving the names of the first original entries
}
\examples{
unique_with_names(c(N = "n", b = "b", A = "n"))
}
\seealso{
Other vectors: 
\code{\link{len_unique}()},
\code{\link{names_values_switch}()},
\code{\link{rle_df}()},
\code{\link{rle_groups}()},
\code{\link{rle_seq}()},
\code{\link{sort_by_names}()},
\code{\link{true_names}()},
\code{\link{unify_vec_preserve_order}()},
\code{\link{vlookup_from_ref}()},
\code{\link{zchar_remove}()}
}
\concept{vectors}
