% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{unit_find}
\alias{unit_find}
\title{Find the row corresponding to a value in \code{unit_conversion}}
\usage{
unit_find(x)
}
\arguments{
\item{x}{\code{chr/num} vector to find}
}
\value{
\code{tbl}
}
\description{
Find the row corresponding to a value in \code{unit_conversion}
}
\examples{
unit_find("B")
}
