% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rounding.R
\name{unit_modify_vec}
\alias{unit_modify_vec}
\title{Modify unit abbreviation, vectorized version}
\usage{
unit_modify_vec(x, unit, outtype, magnitude = magnitude_order(x))
}
\arguments{
\item{x}{\code{num} The maximum number in the dataset}

\item{unit}{\code{chr} Type of units, supported values are: \code{AF, F, $, W, MW}}

\item{outtype}{\code{chr} The type of output to be added. Current possibilities are: \code{begin, end, abbrev, unit_eng}}

\item{magnitude}{\code{num} The order of magnitude for the unit. The highest triplet will be used. See \code{magnitude_triplet}}
}
\value{
\code{chr} updated units
}
\description{
Modify unit abbreviation, vectorized version
}
\examples{
unit_modify(10^7, "AF", "abbrev")
}
\seealso{
unit_modify

Other rounding: 
\code{\link{magnitude_order}()},
\code{\link{magnitude_triplet}()},
\code{\link{num2str}()},
\code{\link{num2str_vec}()},
\code{\link{num_chr_suffi}},
\code{\link{round_to}()},
\code{\link{size}()},
\code{\link{unit_find}()},
\code{\link{unit_modify}()},
\code{\link{unit_string}()}
}
\concept{rounding}
