% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rounding.R
\name{unit_string}
\alias{unit_string}
\alias{unit_string<-}
\title{Extract the units from a string}
\usage{
unit_string(x)

unit_string(x) <- value
}
\arguments{
\item{x}{\code{chr} String(s) to extract units from/assign units to}
}
\value{
\code{chr}
}
\description{
It is assumed that units are encased in parentheses at the end of the string
}
\examples{
unit_string("Elevation (F)")
}
\seealso{
Other rounding: 
\code{\link{magnitude_order}()},
\code{\link{magnitude_triplet}()},
\code{\link{num2str}()},
\code{\link{num2str_vec}()},
\code{\link{num_chr_suffi}},
\code{\link{round_to}()},
\code{\link{size}()},
\code{\link{unit_find}()},
\code{\link{unit_modify}()},
\code{\link{unit_modify_vec}()}

Other rounding: 
\code{\link{magnitude_order}()},
\code{\link{magnitude_triplet}()},
\code{\link{num2str}()},
\code{\link{num2str_vec}()},
\code{\link{num_chr_suffi}},
\code{\link{round_to}()},
\code{\link{size}()},
\code{\link{unit_find}()},
\code{\link{unit_modify}()},
\code{\link{unit_modify_vec}()}
}
\concept{rounding}
