% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environments_namespaces.R
\name{unload_namespaces}
\alias{unload_namespaces}
\title{Unload namespaces prior to package install}
\usage{
unload_namespaces(ns, verbose = FALSE)
}
\arguments{
\item{ns}{\code{(chr)} namespaces to unload}

\item{verboe}{\code{lgl}}
}
\description{
Unload namespaces prior to package install
}
