% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{use_reimport}
\alias{use_reimport}
\title{Add a function to reimports}
\usage{
use_reimport(pkg, fun)
}
\arguments{
\item{pkg}{\code{chr} package name}

\item{fun}{\code{chr} function name}
}
\value{
\code{msg}
}
\description{
Add a function to reimports
}
\seealso{
Other project setup: 
\code{\link{creds_to_renviron}()},
\code{\link{fun_docs_table}()},
\code{\link{ignore_files}()},
\code{\link{install_remote}()},
\code{\link{key_pairs_duplicated}()},
\code{\link{key_pairs_text}()},
\code{\link{use_UU_reimports}()},
\code{\link{write_to_rprofile}()}
}
\concept{project setup}
