% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_helpers.R
\name{week_factor}
\alias{week_factor}
\title{Day of the week as factor/numeric}
\usage{
week_factor(
  x,
  label = FALSE,
  abbr = TRUE,
  week_start = getOption("lubridate.week.start", 7)
)
}
\arguments{
\item{x}{\code{chr/Date/POSIXt} Days of the week as case insensitive character vector, if using abbreviations, they must be of consistent length. Otherwise, a \code{Date} or \code{POSIXt}}
}
\value{
\code{numeric/factor} Depending on the value for \code{label}. If \code{x} is provided, the day of the week as a factor. If \code{x} is not provided, an ordered factor with Monday as 1.
}
\description{
Day of the week as factor/numeric
}
\examples{
week_factor()
week_factor(c('Tu', 'We'))
week_factor(c('Tu', 'We'), label = TRUE)
week_factor(seq(lubridate::floor_date(Sys.time(), "year"), Sys.time(), by = "day"), label = TRUE, abbr = FALSE)
}
\seealso{
Other time: 
\code{\link{duration_print}()},
\code{\link{excel_date}()},
\code{\link{month_factor}()},
\code{\link{season_factor}()},
\code{\link{time_aggregates}},
\code{\link{time_difftimes}},
\code{\link{time_elapsed}()},
\code{\link{time_factor}()},
\code{\link{timespan}()}
}
\concept{time}
