% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{write_dir_fn}
\alias{write_dir_fn}
\title{Write \code{dir} helper function that are robust to dev vs deployed package states}
\usage{
write_dir_fn(outfile = "R/utils_dir_fns.R", overwrite = TRUE)
}
\arguments{
\item{outfile}{\code{chr} path to file to write. Default \emph{R/utils_dir_fns.R}}

\item{overwrite}{\code{lgl} Whether to overwrite the existing file. Default \code{TRUE}}
}
\value{
\code{msg} and a new file
}
\description{
Write \code{dir} helper function that are robust to dev vs deployed package states
}
