% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_io.R
\name{write_lines}
\alias{write_lines}
\title{Write lines at a specific location in a file}
\usage{
write_lines(file, ..., after = "end")
}
\arguments{
\item{file}{\code{chr} path to file}

\item{...}{\code{chr} lines to write}

\item{after}{\code{num/chr} either a line number of "end" to write it at the end}
}
\value{
the resulting file
}
\description{
Write lines at a specific location in a file
}
\seealso{
Other file IO: 
\code{\link{col_types}()},
\code{\link{dep_read}()},
\code{\link{dep_write}()},
\code{\link{dir_fn}()},
\code{\link{dirs}},
\code{\link{ext}()},
\code{\link{file_fn}()},
\code{\link{is_filepath}()},
\code{\link{last_updated}()},
\code{\link{list.files2}()},
\code{\link{make_names}()},
\code{\link{mkpath}()},
\code{\link{move_files_to_folder}()},
\code{\link{needs_update}()},
\code{\link{object_fn}()},
\code{\link{object_write}()},
\code{\link{package_size}()},
\code{\link{write_dir_fn}()}
}
\concept{file IO}
