% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{write_to_rprofile}
\alias{write_to_rprofile}
\title{Write expressions to the \emph{.Rprofile}}
\usage{
write_to_rprofile(..., scope = c("user", "project")[1])
}
\arguments{
\item{...}{\code{exprs}}

\item{scope}{\code{chr} which \emph{.Rprofile} to write to}
}
\value{
\coe{msg}
}
\description{
Write expressions to the \emph{.Rprofile}
}
