% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_useful.R
\name{zchar}
\alias{zchar}
\title{Is zero-length character?}
\usage{
zchar(...)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{lgl}
}
\description{
Is zero-length character?
}
\examples{
zchar("")
zchar(" ")
}
\seealso{
Other character: 
\code{\link{zchar_remove}()}
}
\concept{character}
