% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helpers.R
\name{zchar_remove}
\alias{zchar_remove}
\title{Remove zero length strings (or string with all spaces)}
\usage{
zchar_remove(x)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{chr}
}
\description{
Remove zero length strings (or string with all spaces)
}
\examples{
zchar_remove(c("", "  ", "a"))
}
